% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aggregate.R
\name{aggregateRegionsByGroup}
\alias{aggregateRegionsByGroup}
\title{aggregateRegionsByGroup}
\usage{
aggregateRegionsByGroup(
  ct2,
  group_VAR,
  new_region_VAR = ifelse(length(group_VAR) == 1, group_VAR, "grouped_regions"),
  sep = " ",
  summary_FUN = mean
)
}
\arguments{
\item{ct2}{A ChIPtsne2 object.}

\item{group_VAR}{Attribute name to aggregate regions to.  There will be 1 meta-region per unique entry in \code{group_VAR}. \code{group_VAR} may specify multiple attributes, in which case there will 1 meta-region per combination of entries in all \code{group_VAR}.}

\item{new_region_VAR}{The new region variable of the resulting ChIPtsne2 object. Defaults to \code{group_VAR} for single \code{group_VAR} but then defaults to "grouped_regions" if there are multiple.}

\item{sep}{Works just like \code{sep} parameter for base::paste. Used to combine multiple group_VAR values.}

\item{summary_FUN}{function applied to every column to summarize rows. Default is mean.}
}
\value{
A ChIPtsne2_no_rowRanges object with meta-regions for combinations of \code{group_VAR} values.
}
\description{
aggregateRegionsByGroup
}
\examples{
ct2 = exampleChIPtsne2.with_meta()

ct2.agg1 = aggregateRegionsByGroup(ct2, "peak_MCF10A_CTCF")
rowData(ct2.agg1)
getRegionVariable(ct2.agg1)

ct2.agg2 = aggregateRegionsByGroup(
  ct2,
  c("peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"),
  new_region_VAR = "peak_overlap"
)
rowData(ct2.agg2)
getRegionVariable(ct2.agg2)
}
