% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_centroid_classifier.R
\name{calculateGroupCentroid}
\alias{calculateGroupCentroid}
\title{calculateGroupCentroid}
\usage{
calculateGroupCentroid(ct2, group_VARS, sep = ",")
}
\arguments{
\item{ct2}{\verb{R doc_ct2()}}

\item{group_VARS}{A least 1 categorical attribute in rowData of \code{ct2}. When
multiple attributes are specified, centroids will be calculated for all
combinations of groups.}

\item{sep}{Separator character to use when pasting multiple grouping variables.}
}
\value{
A matrix of centroid profiles per group. Input to
\link{groupRegionsByCentroidDistance}.
}
\description{
calculateGroupCentroid
}
\examples{
ct2 = exampleChIPtsne2()
cents = calculateGroupCentroid(ct2, c("peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"))
cents
ct2 = groupRegionsByCentroidDistance(ct2, cents, group_VAR = "peak_overlap_profile")
rowData(ct2)
}
