% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_alt_constructors.R
\name{ChIPtsne2.history}
\alias{ChIPtsne2.history}
\title{ChIPtsne2.history}
\usage{
ChIPtsne2.history(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}
}
\value{
list of history items
}
\description{
ChIPtsne2.history
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ChIPtsne2.history(ct2)
ct2 = setNameVariable(ct2, "name2")
# the previous operation is now in the history
ChIPtsne2.history(ct2)
}
