% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsneList.R
\name{ChIPtsne2List}
\alias{ChIPtsne2List}
\title{ChIPtsne2List}
\usage{
ChIPtsne2List(...)
}
\arguments{
\item{...}{A number of ChIPtsne2_no_rowRanges or ChIPtsne2 objects.}
}
\description{
ChIPtsne2List
}
