% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_centroid_classifier.R
\name{classify_by_centroid_distances}
\alias{classify_by_centroid_distances}
\title{classify_by_centroid}
\usage{
classify_by_centroid_distances(
  distances,
  centroids,
  ambiguous_value = "ambiguous",
  tolerance = 0
)
}
\arguments{
\item{centroids}{}
}
\description{
classify_by_centroid
}
