% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_normalize_signal.R
\name{calculateSignalCapValue}
\alias{calculateSignalCapValue}
\alias{calculateSignalCapValue,ChIPtsne2_no_rowRanges-method}
\alias{normalizeSignalPerRegion,ChIPtsne2_no_rowRanges-method}
\title{calculateSignalCapValue}
\usage{
calculateSignalCapValue(ct2, signal_cap_VAR = "cap_value", cap_quantile = 0.95)

\S4method{calculateSignalCapValue}{ChIPtsne2_no_rowRanges}(ct2, signal_cap_VAR = "cap_value", cap_quantile = 0.95)

\S4method{normalizeSignalPerRegion}{ChIPtsne2_no_rowRanges}(ct2, group_VARS = NULL)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{signal_cap_VAR}{The attribute name to store results in colData/sample metadata. Will be overwritten if present.}

\item{cap_quantile}{The quantile at which to set cap value. Default is .95.}
}
\value{
A ChIPtsne2_no_rowRanges object. where colData/sample metadata has had \code{signal_cap_VAR} added.
}
\description{
Calculates but does not apply a cap value for each sample. Meant to be run before \link{normalizeSignalCapValue}.
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = calculateSignalCapValue(ct2)
ct2 = calculateSignalCapValue(ct2, signal_cap_VAR = "higher_cap_value", cap_quantile = .98)
colData(ct2)
}
