% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce.R
\name{dimReduceTSNE}
\alias{dimReduceTSNE}
\alias{dimReduceTSNE,ChIPtsne2_no_rowRanges-method}
\alias{dimReduceUMAP}
\alias{dimReduceUMAP,ChIPtsne2_no_rowRanges-method}
\alias{dimReducePCA}
\alias{dimReducePCA,ChIPtsne2_no_rowRanges-method}
\title{dimReduceTSNE}
\usage{
dimReduceTSNE(ct2, perplexity = 50, ...)

\S4method{dimReduceTSNE}{ChIPtsne2_no_rowRanges}(ct2, perplexity = 50, ...)

dimReduceUMAP(ct2, config = umap::umap.defaults, ...)

\S4method{dimReduceUMAP}{ChIPtsne2_no_rowRanges}(ct2, config = umap::umap.defaults, ...)

dimReducePCA(ct2)

\S4method{dimReducePCA}{ChIPtsne2_no_rowRanges}(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{perplexity}{Perplexity value for calculation of tSNE. Will be automatically reduced if too large. Default is 50.}

\item{...}{Additional arguments passed to \link[umap:umap]{umap::umap}}

\item{config}{Configuration used by umap package. Default is \link[umap:umap.defaults]{umap::umap.defaults}}
}
\value{
A ChIPtsne2_no_rowRanges object. with tSNE coordinates added to region meta data.

A ChIPtsne2_no_rowRanges object. with UMAP coordinates added to region meta data.

A ChIPtsne2_no_rowRanges object. with PCA coordinates added to region meta data.
}
\description{
dimReduceTSNE

dimReduceUMAP

dimReducePCA
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = dimReduceTSNE(ct2)
ct2
ct2 = exampleChIPtsne2.with_meta()
ct2 = dimReduceUMAP(ct2)
ct2
ct2 = exampleChIPtsne2.with_meta()
ct2 = dimReducePCA(ct2)
ct2
}
