% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_no_rowRanges.R
\name{cbind,ChIPtsne2_no_rowRanges-method}
\alias{cbind,ChIPtsne2_no_rowRanges-method}
\alias{rbind,ChIPtsne2_no_rowRanges-method}
\alias{rbind}
\alias{cbind}
\title{cbind-ChIPtsne2_no_rowRanges}
\usage{
\S4method{cbind}{ChIPtsne2_no_rowRanges}(..., deparse.level = 1)

\S4method{rbind}{ChIPtsne2_no_rowRanges}(..., deparse.level = 1)
}
\arguments{
\item{...}{A single A ChIPtsne2_no_rowRanges object.}

\item{deparse.level}{Not used.}

\item{ChIPtsne2_no_rowRanges}{A ChIPtsne2_no_rowRanges object.}
}
\value{
a ChIPtsne2 object of concatenated columns/samples of all items in input ChIPtsne2List

a ChIPtsne2 object of concatenated rows/regions of all items in input ChIPtsne2List
}
\description{
cbind-ChIPtsne2_no_rowRanges

rbind-ChIPtsne2_no_rowRanges
}
\examples{
ct2.left = exampleChIPtsne2()
ct2.right = exampleChIPtsne2()
colnames(ct2.right) = paste0("right_", colnames(ct2.right))
ct2 = cbind(ct2.left, ct2.right)
dim(ct2)
colnames(ct2)

rowData(ct2.left)$tx = seq(100)
rowData(ct2.right)$tx = 1
colData(ct2.left)$tmp = "A"
ct2 = cbind(ct2.left, ct2.right)
ct2_a = exampleChIPtsne2()
ct2_b = exampleChIPtsne2()
# duplicated rownames are not allowed so we need to modify before rbind
rownames(ct2_b) = paste0("b_", rownames(ct2_b))
ct2_rbind = rbind(ct2_a, ct2_b)
rownames(ct2_rbind)

rowData(ct2_a)$tx = seq(100)
colData(ct2_a)$tmp = "A"
colData(ct2_b)$tmp = "B"
ct2 = rbind(ct2_a, ct2_b)

rowData(ct2)
colData(ct2)
}
