% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_grouping.R
\name{groupRegionsBySignalCluster}
\alias{groupRegionsBySignalCluster}
\alias{groupRegionsBySignalCluster,ChIPtsne2_no_rowRanges-method}
\title{groupRegionsBySignalCluster}
\usage{
groupRegionsBySignalCluster(
  ct2,
  group_VAR = "cluster_id",
  n_clusters = 6,
  iter.max = 30
)

\S4method{groupRegionsBySignalCluster}{ChIPtsne2_no_rowRanges}(
  ct2,
  group_VAR = "cluster_id",
  n_clusters = 6,
  iter.max = 30
)
}
\arguments{
\item{ct2}{A ChIPtsne2 object.}

\item{group_VAR}{Attribute name to add to rowData for storing new row/region grouping assignment. Will be overwritten if it exists.}

\item{n_clusters}{Number of clusters specified for k-means.}

\item{iter.max}{Number of max iterations to allow for k-means. Default is 30. Passed to \link[stats:kmeans]{stats::kmeans}.}
}
\value{
Updated ChIPtsne2 object with group assignment added to rowData using specified group_VAR.
}
\description{
groupRegionsBySignalCluster
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = groupRegionsBySignalCluster(ct2)
ct2 = groupRegionsBySignalCluster(ct2)
}
