% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_normalize_signal.R
\name{normalizeSignalRPM}
\alias{normalizeSignalRPM}
\alias{normalizeSignalRPM,ChIPtsne2_no_rowRanges-method}
\title{normalizeSignalRPM}
\usage{
normalizeSignalRPM(
  ct2,
  mapped_reads_data = NULL,
  mapped_reads_VAR = "mapped_reads"
)

\S4method{normalizeSignalRPM}{ChIPtsne2_no_rowRanges}(
  ct2,
  mapped_reads_data = NULL,
  mapped_reads_VAR = "mapped_reads"
)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{mapped_reads_data}{Optionally, a data.frame containing \code{ct2}'s name
variable and specified \code{mapped_reads_VAR}. Otherwise colData/sample
metadata will be used.}

\item{mapped_reads_VAR}{Variable name in colData/sample metadata or
\code{mapped_reads_data} if provided.}
}
\value{
A ChIPtsne2_no_rowRanges object. where value_VAR has been normalized according to
\code{mapped_reads_VAR}
}
\description{
normalizeSignalRPM
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
# we're going to spoof in mapped reads here
# but if you have bam files you can use [get_mapped_reads].
colData(ct2)$mapped_reads = c(3e6, 2e6, 1e6)
plotSignalLinePlot(ct2)
ct2.rpm = normalizeSignalRPM(ct2)
# RPM normalization corrects for depth, reversing magnitude trend in this exampl data.
plotSignalLinePlot(ct2.rpm)
}
