% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.R
\name{rowRanges<-,ChIPtsne2,NULL-method}
\alias{rowRanges<-,ChIPtsne2,NULL-method}
\title{rowRanges accessor of ChIPtsne2}
\usage{
\S4method{rowRanges}{ChIPtsne2,NULL}(x, ...) <- value
}
\arguments{
\item{x}{A ChIPtsne2 object.}

\item{...}{not used}

\item{value}{Only NULL is allowed.}
}
\value{
A ChIPtsne2_no_rowRanges object. that is the same as input ChIPtsne2 object but with rowRanges removed.
}
\description{
rowRanges accessor of ChIPtsne2
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
class(ct2)
rowRanges(ct2) = NULL
#class has changed to ChIPtsne2_no_rowRanges
class(ct2)
}
