% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_converters.R
\name{getTidyProfile}
\alias{getTidyProfile}
\alias{getTidyProfile,ChIPtsne2_no_rowRanges-method}
\title{getTidyProfile}
\usage{
getTidyProfile(ct2, meta_VARS = NULL)

\S4method{getTidyProfile}{ChIPtsne2_no_rowRanges}(ct2, meta_VARS = NULL)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{meta_VARS}{Attributes from either sample metadata/colData or region metadata/rowData to include.}
}
\value{
A tidy data represenation of profiles stored in \code{ct2}.
}
\description{
getTidyProfile
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
getTidyProfile(ct2)
}
