% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce.R
\name{dimReducePCA}
\alias{dimReducePCA}
\title{dimReducePCA}
\usage{
dimReducePCA(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}
}
\value{
A ChIPtsne2_no_rowRanges object. with PCA coordinates added to region meta data.
}
\description{
dimReducePCA
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = dimReducePCA(ct2)
ct2
}
