% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce.R
\name{dimReduceTSNE}
\alias{dimReduceTSNE}
\title{dimReduceTSNE}
\usage{
dimReduceTSNE(ct2, perplexity = 50, ...)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{perplexity}{Perplexity value for calculation of tSNE. Will be automatically reduced if too large. Default is 50.}

\item{...}{Additional arguments passed to \link[Rtsne:Rtsne]{Rtsne::Rtsne}}
}
\value{
A ChIPtsne2_no_rowRanges object. with tSNE coordinates added to region meta data.
}
\description{
dimReduceTSNE
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = dimReduceTSNE(ct2)
ct2
}
