% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce.R
\name{dimReduceUMAP}
\alias{dimReduceUMAP}
\title{dimReduceUMAP}
\usage{
dimReduceUMAP(ct2, config = umap::umap.defaults, ...)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{config}{Configuration used by umap package. Default is \link[umap:umap.defaults]{umap::umap.defaults}}

\item{...}{Additional arguments passed to \link[umap:umap]{umap::umap}}
}
\value{
A ChIPtsne2_no_rowRanges object. with UMAP coordinates added to region meta data.
}
\description{
dimReduceUMAP
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = dimReduceUMAP(ct2)
ct2
}
