% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_center_signal.R
\name{.centerSignal}
\alias{.centerSignal}
\title{.centerSignal}
\usage{
.centerSignal(ct2)
}
\arguments{
\item{ct}{}
}
\value{
A chiptsne2 object updated to reflect centering procecure.
}
\description{
.centerSignal
}
\examples{
library(tidyverse)
query_gr = seqsetvis::CTCF_in_10a_overlaps_gr
query_gr = seqsetvis::prepare_fetch_GRanges_width(query_gr, win_size = 50)
prof_dt = seqsetvis::CTCF_in_10a_profiles_dt
meta_dt = prof_dt \%>\%
  select(sample) \%>\%
  unique \%>\%
  separate(sample, c("cell", "mark"), remove = FALSE)
ct2 = ChIPtsne2.from_tidy(prof_dt, query_gr, sample_metadata = meta_dt)
debug(chiptsne2:::.centerSignal)
chiptsne2:::.centerSignal(ct2)
}
