% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_center_signal.R
\name{.centerSignalGRanges}
\alias{.centerSignalGRanges}
\title{.centerSignal}
\usage{
.centerSignalGRanges(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne2 object}
}
\value{
A chiptsne2 object updated to reflect centering procedure. Width will be the same as original but this requires a second fetch.
}
\description{
.centerSignal
}
\examples{
library(tidyverse)
query_gr = seqsetvis::CTCF_in_10a_overlaps_gr
query_gr = seqsetvis::prepare_fetch_GRanges_width(query_gr, win_size = 50)
prof_dt = seqsetvis::CTCF_in_10a_profiles_dt
meta_dt = prof_dt \%>\%
  select(sample) \%>\%
  unique \%>\%
  separate(sample, c("cell", "mark"), remove = FALSE)
ct2 = ChIPtsne2.from_tidy(prof_dt, query_gr, sample_metadata = meta_dt)
chiptsne2:::.centerSignalGRanges(ct2)
}
