% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce_plot_points.R
\name{.enforce_extra_VARS}
\alias{.enforce_extra_VARS}
\title{.enforce_extra_VARS}
\usage{
.enforce_extra_VARS(ct2, df, extra_VARS, expected_missing = NULL)
}
\arguments{
\item{ct2}{chiptsne2 object}

\item{df}{current data.frame to add extra_VARS to}

\item{extra_VARS}{extra variables that must be in metadata or in df already}

\item{expected_missing}{it's ok if these are in extra_VARS and not present in
df or meta data}
}
\value{
df with extra_VARS added from meta data.
}
\description{
.enforce_extra_VARS
}
