% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_flip_signal.R
\name{.flip_regions}
\alias{.flip_regions}
\title{.flip_regions}
\usage{
.flip_regions(ct2, to_flip)
}
\arguments{
\item{ct2}{valid chiptsne2 object}

\item{to_flip}{selector for rownames in ct2. accepts logical or character vector.}
}
\value{
ChIPtsne2 object with regions flipped according to to_flip.
}
\description{
.flip_regions
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = groupRegionsBySignalCluster(ct2)
ct2 = sortRegions(ct2, sort_strategy = "left", group_VAR = "cluster_id")
p1 = plotSignalHeatmap(ct2, sort_strategy = "none", group_VARS = "cluster_id")

to_flip = rowData(ct2)$cluster_id == 1
ct2_flipped = .flip_regions(ct2, to_flip)
p2 = plotSignalHeatmap(ct2_flipped, sort_strategy = "none", group_VARS = "cluster_id")
cowplot::plot_grid(p1, p2)

ct2_flipped2 = .flip_regions(ct2, which(to_flip))
p3 = plotSignalHeatmap(ct2_flipped2, sort_strategy = "none", group_VARS = "cluster_id")
cowplot::plot_grid(p1, p3)

ct2_flipped3 = .flip_regions(ct2, rownames(ct2)[to_flip])
p4 = plotSignalHeatmap(ct2_flipped3, sort_strategy = "none", group_VARS = "cluster_id")
cowplot::plot_grid(p1, p4)
}
