% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_flip_signal.R
\name{.flipProfilesToMatch}
\alias{.flipProfilesToMatch}
\title{.centerSignal}
\usage{
.flipProfilesToMatch(ct2, highest_on_right = FALSE)
}
\arguments{
\item{ct2}{A ChIPtsne2 object}
}
\value{
A chiptsne2 object updated to reflect centering procedure. Width will be the same as original but this requires a second fetch.
}
\description{
.centerSignal
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2_left = flipProfilesToMatch(ct2)
ct2_right = flipProfilesToMatch(ct2, highest_on_right = TRUE)

prof_original = getTidyProfile(ct2)
prof_original$group = "original"

prof_left = getTidyProfile(ct2_left)
prof_left$group = "flip_left"

prof_right = getTidyProfile(ct2_right)
prof_right$group = "flip_right"

prof_dt = rbind(prof_left, prof_right, prof_original)
prof_dt = prof_dt[, .(y = mean(y)), .(x, group)]
ggplot(prof_dt, aes(x = x, y = y, color = group)) + geom_path()
}
