% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_grouping.R
\name{.groupRegionsByOverlap}
\alias{.groupRegionsByOverlap}
\title{Title}
\usage{
.groupRegionsByOverlap(ct2, gr_list, group_VAR = "overlap_id")
}
\arguments{
\item{group_VAR}{}
}
\description{
Title
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
peak_grs = seqsetvis::CTCF_in_10a_narrowPeak_grs
ct2.olap = groupRegionsByOverlap(ct2, peak_grs[1:2], group_VAR = "10A_AT1_overlap")
}
