% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_harmonize.R
\name{.harmonize_cbind}
\alias{.harmonize_cbind}
\title{.harmonize_cbind}
\usage{
.harmonize_cbind(
  to_harmonize,
  reference,
  allow_subset_match = FALSE,
  allow_missing_colData = TRUE
)
}
\arguments{
\item{to_harmonize}{ChIPtsne2 object to harmonize in preparation to cbind.}

\item{reference}{ChIPtsne2 or rowData to use as a reference when harmonizing.}

\item{allow_subset_match}{If TRUE, rows of to_harmonize may be a subset/superset of rows of reference. Default is FALSE.}

\item{allow_missing_colData}{If TRUE, colData may have additional attributes}
}
\value{
ChIPtsne2 object derived from to_harmonize that will be compatible to cbind with reference. If reference must be a modified further, a warning will be generated.
}
\description{
.harmonize_cbind
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2.by_cell = split(ct2, "cell")

ct2_ref = ct2.by_cell$MCF10A
ct2_harm = ct2.by_cell$MCF10AT1
colData(ct2_harm)$col_test = "test"
rowData(ct2_harm)$row_test = "test"
rowData(ct2_harm)$peak_MCF10CA1_CTCF = TRUE
ct2_harm[seq(nrow(ct2_harm), 1),]

ct2_harm2 = .harmonize_cbind(ct2_harm, ct2_ref)

ct2_new = cbind(ct2_ref, ct2_harm2[seq(nrow(ct2_harm), 1),])
ct2_new2 = cbind(ct2_harm2[seq(nrow(ct2_harm), 1),], ct2_ref)

rowData(ct2_new)
rowData(ct2_new2)
}
