% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{.message_list}
\alias{.message_list}
\title{message_list}
\usage{
.message_list(x)
}
\arguments{
\item{x}{named list}
}
\value{
Outputs list as a message and returns the character invisibly.
}
\description{
output message of list contents
}
\details{
code based on  \link[base:print]{base::print.listof}.
}
\examples{
x = list(a = 1:3, b = 2:6)
.message_list = chiptsne2:::.message_list
.message_list(x)

y = list(1:3, 2:6)
.message_list(y)
}
