% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{.parse_config_header}
\alias{.parse_config_header}
\title{.parse_config_header}
\usage{
.parse_config_header(
  f,
  valid_config_VARS = c("main_dir", "data_dir", "file_prefix", "view_size",
    "window_size", "read_mode", "fetch_options", "is_null", "name_VAR"),
  allowed_deprecated_VARS = c("color_by", "color_mapping", "run_by", "to_run",
    "to_run_reference")
)
}
\arguments{
\item{f}{a config file}

\item{valid_config_VARS}{supported variables to attempt to extract}

\item{allowed_deprecated_VARS}{These are allowed but ignored. These are ssvQC
config variables that are not relevant to chiptsne2. This allows ssvQC
configs to be reused for chiptsne2.}
}
\value{
A named list containing configuration options mapped to values.
}
\description{
.parse_config_header
}
\examples{
bam_config_file = exampleBamConfigFile()
chiptsne2:::.parse_config_header(bam_config_file)
}
