% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce_plot_summary_profiles.R
\name{.plotDimReduceSummaryProfiles}
\alias{.plotDimReduceSummaryProfiles}
\title{plotDimReduceSummaryProfiles}
\usage{
.plotDimReduceSummaryProfiles(
  ct2,
  color_VAR = ct2@name_VAR,
  x_bins = min(nrow(ct2)^0.5, 10),
  y_bins = x_bins,
  extra_vars = character(),
  xrng = NULL,
  yrng = NULL,
  value_limits = c(0, 1),
  ma_size = 2,
  n_splines = 10,
  p = NULL,
  N_floor = 0,
  N_ceiling = NULL,
  min_size = 0.3,
  return_data = FALSE
)
}
\arguments{
\item{return_data}{}
}
\description{
plotDimReduceSummaryProfiles
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = dimReducePCA(ct2)
plotDimReduceSummaryProfiles(ct2)
plotDimReduceSummaryProfiles(ct2, value_limits = c(0, 50))
plotDimReduceSummaryProfiles(ct2, value_limits = c(0, 200))
plotDimReduceSummaryProfiles(ct2, value_limits = c(0, 500))
}
