% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_history.R
\name{.rerun_history}
\alias{.rerun_history}
\title{rerun_history}
\usage{
.rerun_history(ct2, history_source)
}
\arguments{
\item{ct2}{A ChIPtsne2 object to apply history too.}

\item{history_source}{list of history items or a ChIPtsne2 object to pull history from.}
}
\value{
Input ct2 with all functions from history_source applied.
}
\description{
rerun_history
}
\examples{
library(tidyverse)
query_gr = seqsetvis::CTCF_in_10a_overlaps_gr
query_gr = seqsetvis::prepare_fetch_GRanges_width(query_gr, win_size = 50)
prof_dt = seqsetvis::CTCF_in_10a_profiles_dt
meta_dt = prof_dt \%>\%
  select(sample) \%>\%
  unique \%>\%
  separate(sample, c("cell", "mark"), remove = FALSE)
ct2 = ChIPtsne2.from_tidy(prof_dt, query_gr, sample_metadata = meta_dt)
ct2.c = chiptsne2:::.centerProfilesAndTrim(ct2, view_size = 500)
rerun_history(ct2, ct2.c)
rerun_history(ct2, ChIPtsne2.history(ct2.c))
}
