% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{.shiny_secondary_plot}
\alias{.shiny_secondary_plot}
\title{.shiny_secondary_plot}
\usage{
.shiny_secondary_plot(plot_ct2, tx_df, plotType, selAttribute, selIds)
}
\arguments{
\item{plot_ct2}{A ChIPtsne2_no_rowRanges object. to plot}

\item{tx_df}{dim reduction data.frame with region ids, tx, ty, and other metadata}

\item{plotType}{Allowed types: line_color_selection, line_color_annotation, heatmap_cluster_selection, heatmap_cluster_annotation}

\item{selAttribute}{Grouping attribute to facet y by}

\item{selIds}{ids selection. used for color.}
}
\description{
the secondary plot is meant to compare selections in dim reduces space and
possibly with group context.
}
\examples{

plot_ct2 = exampleChIPtsne2.with_meta()
plot_ct2 = dimReduceTSNE(plot_ct2)
tx_df = getRegionMetaData(plot_ct2)
plotType = "line_color_selection"
selAttribute = "peak_MCF10A_CTCF"
selIds = character()
selIds = rownames(plot_ct2)[1:10]

chiptsne2:::.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "line_color_selection",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = rownames(plot_ct2)[1:10]
)
chiptsne2:::.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "line_color_selection",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = character()
)
chiptsne2:::.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "line_color_annotation",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = rownames(plot_ct2)[1:10]
)
chiptsne2:::.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "line_color_annotation",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = character()
)
chiptsne2:::.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "heatmap_cluster_selection",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = rownames(plot_ct2)[1:10]
)
chiptsne2:::.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "heatmap_cluster_selection",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = character()
)
chiptsne2:::.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "heatmap_cluster_annotation",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = rownames(plot_ct2)[1:10]
)
chiptsne2:::.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "heatmap_cluster_annotation",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = character()
)

}
