% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FetchConfig.R
\name{.show_Config}
\alias{.show_Config}
\title{.show_Config}
\usage{
.show_Config(qc)
}
\arguments{
\item{qc}{A FetchConfig object}
}
\description{
used by show
}
\examples{
bam_config_file = exampleBamConfigFile()
fetch_config = FetchConfig.load_config(bam_config_file)
fetch_config
}
