% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce.R
\name{.umap_from_profile_mat}
\alias{.umap_from_profile_mat}
\title{.umap_from_profile_mat}
\usage{
.umap_from_profile_mat(prof_mat, config = umap::umap.defaults, ...)
}
\arguments{
\item{prof_mat}{matrix of profile data}

\item{config}{umap.config object, uses umap::umap.defaults by default}

\item{...}{passed to umap::umap}
}
\value{
data.frame of reduced dimensions by UMAP
}
\description{
.umap_from_profile_mat
}
\examples{
ct2 = exampleChIPtsne2()
prof_mat = rowToRowMat(ct2)
.umap_from_profile_mat(prof_mat)
}
