% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{exampleBam_data.frame}
\alias{exampleBam_data.frame}
\title{exampleBam_data.frame}
\usage{
exampleBam_data.frame()
}
\value{
data.frame with example bam files and metadata.
}
\description{
exampleBam_data.frame
}
\examples{
exampleBam_data.frame()
}
