% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{exampleBamConfigFile}
\alias{exampleBamConfigFile}
\title{exampleBamConfigFile}
\usage{
exampleBamConfigFile()
}
\value{
character file path to example bam config file.
}
\description{
exampleBamConfigFile
}
\examples{
exampleBamConfigFile()
}
