% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{exampleBamFiles}
\alias{exampleBamFiles}
\title{exampleBamFiles}
\usage{
exampleBamFiles()
}
\value{
character file paths to example bam files.
}
\description{
exampleBamFiles
}
\examples{
exampleBamFiles()
}
