% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{exampleBigWig_data.frame}
\alias{exampleBigWig_data.frame}
\title{exampleBigWig_data.frame}
\usage{
exampleBigWig_data.frame()
}
\value{
data.frame with example bigWig files and metadata.
}
\description{
exampleBigWig_data.frame
}
\examples{
exampleBigWig_data.frame()
}
