% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{exampleBigWigConfigFile}
\alias{exampleBigWigConfigFile}
\title{exampleBigWigConfigFile}
\usage{
exampleBigWigConfigFile()
}
\value{
character file path to example bigWig config file.
}
\description{
exampleBigWigConfigFile
}
\examples{
exampleBigWigConfigFile()
}
