% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{exampleBigWigFiles}
\alias{exampleBigWigFiles}
\title{exampleBigWigFiles}
\usage{
exampleBigWigFiles()
}
\value{
character file paths to example bigWig files.
}
\description{
exampleBigWigFiles
}
\examples{
exampleBigWigFiles()
}
