% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{exampleChIPtsne2.with_meta}
\alias{exampleChIPtsne2.with_meta}
\title{exampleChIPtsne2.with_meta}
\usage{
exampleChIPtsne2.with_meta()
}
\value{
ChIPtsne object for testing, includes meta data
}
\description{
exampleChIPtsne2.with_meta
}
\examples{
exampleChIPtsne2.with_meta()
}
