% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{exampleDataPaths}
\alias{exampleDataPaths}
\title{exampleDataPaths}
\usage{
exampleDataPaths()
}
\value{
data.frame with URL paths to FE bigwigs and narrowPeak files.
}
\description{
exampleDataPaths
}
\examples{
exampleDataPaths()
}
