% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FetchConfig-class}
\alias{FetchConfig-class}
\alias{.FetchConfig}
\title{FetchConfig-class}
\description{
FetchConfig-class
}
\section{Slots}{

\describe{
\item{\code{view_size}}{Consistent size to use when viewing assessment regions. Uses
3kb as default.}

\item{\code{read_mode}}{Read mode of signal data, one of bam_SE, bam_PE, or bigwig.
Use CT_READ_MODES$.}

\item{\code{fetch_options}}{Named list of additional arguments to pass to signal
fetch function.}
}}

