% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FetchConfig.R
\name{FetchConfig.from_files}
\alias{FetchConfig.from_files}
\title{FetchConfig.from_files}
\usage{
FetchConfig.from_files(
  file_paths,
  group_names = NULL,
  name_VAR = "name",
  view_size = 3000,
  window_size = getOption("CT_WINDOW_SIZE", 200),
  read_mode = NULL,
  fetch_options = list()
)
}
\arguments{
\item{file_paths}{character paths to files}

\item{group_names}{vector of group names to assign from according to groups}

\item{name_VAR}{Name variable.}

\item{view_size}{view size in bp to apply. Defaults to 3000. Set to NULL, NA, or -1 to suppress application of view_size.}

\item{window_size}{The window size used when fetching signal. Lower values
increase resolution but also RAM usage. Default is 200 bp.}

\item{read_mode}{Read mode of signal data, one of bam_SE, bam_PE, or bigwig.}

\item{fetch_options}{Named list of additional arguments to pass to signal
fetch function.}
}
\value{
a FetchConfig object
}
\description{
FetchConfig.from_files
}
\examples{
bam_files = exampleBamFiles()
FetchConfig.from_files(bam_files)
}
