% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FetchConfig.R
\name{FetchConfig.load_config}
\alias{FetchConfig.load_config}
\title{FetchConfig.load_config}
\usage{
FetchConfig.load_config(signal_config_file, name_VAR = NULL)
}
\arguments{
\item{signal_config_file}{Configuration file for signal data.}

\item{name_VAR}{Name variable. Must be in body of config file. May also be
defined in header. Setting it here will override specification in config
file.}
}
\value{
A FetchConfig object
}
\description{
Load a FetchConfig from a file.
}
\details{
This file consists of a header, with
configuration parameters stored on lines starting with #CFG and a csv body
specifying file paths and metadata.

Generally you would want to first create
a FetchConfig using \link{FetchConfig} or \link{FetchConfig.from_files} then save it
with \link{FetchConfig.save_config}, rather than try to create a config file from scratch.
}
\examples{
bam_config_file = exampleBamConfigFile()
bam_config = FetchConfig.load_config(bam_config_file)

}
