% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FetchConfig.R
\name{FetchConfig.null}
\alias{FetchConfig.null}
\title{FetchConfig null placeholder}
\usage{
FetchConfig.null()
}
\value{
A null/empty FetchConfig object
}
\description{
FetchConfig null placeholder
}
\examples{
# example code
FetchConfig.null()
}
