% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FetchConfig.R
\name{FetchConfig}
\alias{FetchConfig}
\alias{names,FetchConfig-method}
\alias{$,FetchConfig-method}
\alias{$<-,FetchConfig-method}
\title{FetchConfig}
\usage{
FetchConfig(
  config_df,
  read_mode = NULL,
  view_size = 3000,
  window_size = 200,
  fetch_options = list(),
  is_null = FALSE,
  name_VAR = "name"
)

\S4method{names}{FetchConfig}(x)

\S4method{$}{FetchConfig}(x, name)

\S4method{$}{FetchConfig}(x, name) <- value
}
\arguments{
\item{config_df}{A data.frame containing configuration information for signal
(bam or bigwig) files. Must contain a "file" attribute and the attribute
specified by \code{name_VAR}. and color_by.}

\item{read_mode}{Read mode of signal data, one of bam_SE, bam_PE, or bigwig.}

\item{view_size}{Consistent size to use when viewing assessment regions. Uses
3kb as default.}

\item{window_size}{The window size used when fetching signal. Lower values
increase resolution but also RAM usage. Default is 200 bp.}

\item{fetch_options}{Named list of additional arguments to pass to signal
fetch function.}

\item{is_null}{If TRUE, this FetchConfig is considered null/empty.}

\item{name_VAR}{Name variable. Must be in \code{config_df}.}

\item{x}{A FetchConfig object}

\item{name}{Name of slot to be replaced/retrieved}

\item{value}{New value for slot.}
}
\value{
A FetchConfig object
}
\description{
This is the primary method for creating new FetchConfig objects.
}
\details{
Consider loading FetchConfig from a file with \link{FetchConfig.load_config} or
using \link{FetchConfig.from_files} to quickly create a basic FetchConfig from a
vector of files.
}
\examples{
#loading a config from a file
bam_config_file = exampleBamConfigFile()
bam_config = FetchConfig.load_config(bam_config_file)
#FetchConfig.save_config(bam_config, "bam_config.csv")

bigwig_config_file = exampleBigWigConfigFile()
bigwig_config = FetchConfig.load_config(bigwig_config_file)
#FetchConfig.save_config(bigwig_config, "bigwig_config.csv")

#creating config from a new data.frame
bam_config_df = exampleBam_data.frame()
sig_conf = FetchConfig(bam_config_df)

bigwig_config_df = exampleBigWig_data.frame()
sig_conf.bw = FetchConfig(bigwig_config_df)

cfg.null = FetchConfig.null()
isFetchConfigNull(cfg.null)

#creating a config from simple file paths

bam_files = exampleBamFiles()
FetchConfig.from_files(bam_files)

FetchConfig.from_files(bam_files,
  group_names = c(
    "MCF10A_CTCF", "MCF10A_input",
    "MCF10AT1_CTCF", "MCF10AT1_input",
    "MCF10CA1a_CTCF", "MCF10CA1a_input")
)

#Using a config to fetch
bam_config_file = exampleBamConfigFile()
fetch_config = FetchConfig.load_config(bam_config_file)

query_gr = seqsetvis::CTCF_in_10a_overlaps_gr
runFetchAtRegions(fetch_config, query_gr)

}
