% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FetchConfig.R
\name{FetchConfig.save_config}
\alias{FetchConfig.save_config}
\title{FetchConfig.save_config}
\usage{
FetchConfig.save_config(object, file)
}
\arguments{
\item{object}{FetchConfig object to save.}

\item{file}{File path to save FetchConfig object too.}
}
\value{
Invisibly returns path to saved config file.
}
\description{
FetchConfig.save_config
}
\examples{
bam_config_file = exampleBamConfigFile()
bam_config = FetchConfig.load_config(bam_config_file)
#FetchConfig.save_config(bam_config, "bam_config.csv")
}
