% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{get_mapped_reads}
\alias{get_mapped_reads}
\title{get_mapped_reads}
\usage{
get_mapped_reads(bam_files)
}
\arguments{
\item{bam_files}{A bam file.  Matching .bai file must exist.}
}
\value{
The number of mapped reads in bam file.
}
\description{
get_mapped_reads
}
\examples{
bam_files = exampleBamFiles()
get_mapped_reads(bam_files)
}
