% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_getset.R
\name{getMaxValueData,ChIPtsne2_no_rowRanges-method}
\alias{getMaxValueData,ChIPtsne2_no_rowRanges-method}
\title{getMaxValueData}
\usage{
\S4method{getMaxValueData}{ChIPtsne2_no_rowRanges}(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne object}
}
\value{
tidy data.frame with full sample meta data and maximum value per region.
}
\description{
getMaxValueData
}
\examples{
ct2 = exampleChIPtsne2()
getMaxValueData(ct2)
}
