% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_getset.R
\name{getRegionMetaData,ChIPtsne2_no_rowRanges-method}
\alias{getRegionMetaData,ChIPtsne2_no_rowRanges-method}
\title{getRegionMetaData}
\usage{
\S4method{getRegionMetaData}{ChIPtsne2_no_rowRanges}(ct2, select_VARS = NULL, include_value_max = FALSE)
}
\arguments{
\item{ct2}{A ChIPtsne object}

\item{select_VARS}{character vector of variables to select from region
metadata. Default of NULL will select all available region metadata
variables.}

\item{include_value_max}{If TRUE, values from max assay slot will be included per region. Default is FALSE.}
}
\value{
data.frame with region meta data, similar to rowRanges but suitable
for tidyverse operations.
}
\description{
getRegionMetaData
}
\examples{
ct2 = exampleChIPtsne2()
getRegionMetaData(ct2)
getRegionMetaData(ct2, c("peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"))
getRegionMetaData(ct2, c("peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"), include_value_max = TRUE)
}
