% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.R
\name{getRegionMetaData}
\alias{getRegionMetaData}
\title{getRegionMetaData}
\usage{
getRegionMetaData(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne object}
}
\value{
data.frame with region meta data, similar to rowRanges but suitable for tidyverse operations.
}
\description{
getRegionMetaData
}
\examples{
ct2 = exampleChIPtsne2()
getRegionMetaData(ct2)
}
