% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.R
\name{getSampleMetaData}
\alias{getSampleMetaData}
\title{getSampleMetaData}
\usage{
getSampleMetaData(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne object}
}
\value{
data.frame with sample meta data, similar to colData but suitable for tidyverse operations.
}
\description{
getSampleMetaData
}
\examples{
ct2 = exampleChIPtsne2()
getSampleMetaData(ct2)
}
