% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_grouping.R
\name{.groupRegionsBySignalCluster}
\alias{.groupRegionsBySignalCluster}
\title{groupRegionsBySignalCluster}
\usage{
.groupRegionsBySignalCluster(ct2, group_VAR = "cluster_id", n_clusters = 6)
}
\arguments{
\item{n_clusters}{}
}
\description{
groupRegionsBySignalCluster
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = groupRegionsBySignalCluster(ct2)
ct2 = groupRegionsBySignalCluster(ct2)
}
