% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce.R
\name{hasDimReduce}
\alias{hasDimReduce}
\title{hasDimReduce}
\usage{
hasDimReduce(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}
}
\value{
TRUE if input ct2 has coordinates from dimensional reduction data
}
\description{
hasDimReduce
}
\examples{
ct2 = exampleChIPtsne2()
hasDimReduce(ct2)
ct2 = dimReduceTSNE(ct2, perplexity = 25)
hasDimReduce(ct2)
}
