% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FetchConfig.R
\name{isFetchConfigNull}
\alias{isFetchConfigNull}
\title{isFetchConfigNull}
\usage{
isFetchConfigNull(fetch_config)
}
\arguments{
\item{fetch_config}{A FetchConfig object}
}
\value{
TRUE if object is null placeholder
}
\description{
isFetchConfigNull
}
