% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsneList.R
\name{lapply,ChIPtsne2List-method}
\alias{lapply,ChIPtsne2List-method}
\title{lapply-ChIPtsne2List}
\usage{
\S4method{lapply}{ChIPtsne2List}(X, FUN, ...)
}
\arguments{
\item{X}{A ChIPtsne2List object.}

\item{FUN}{Function to apply to each ChIPtsne2 object.}

\item{...}{Passed to lapply as additional arguments for FUN.}
}
\value{
A ChIPtsne2List if FUN returns a ChIPtsne2 object. Otherwise a simple list.
}
\description{
Ensure when you \link[base:lapply]{base::lapply} on a ChIPtsne2List that a ChIPtsne2List object is returned if FUN returns a ChIPtsne2 object.
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2.by_peak = split(ct2, "peak_MCF10AT1_CTCF")
getNameVariable(ct2.by_peak)
ct2.by_peak = lapply(ct2.by_peak, setNameVariable, "name")
getNameVariable(ct2.by_peak)
}
