% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_normalize_signal.R
\name{normalizeSignalPerRegion}
\alias{normalizeSignalPerRegion}
\title{normalizeSignalPerRegion}
\usage{
normalizeSignalPerRegion(ct2, group_VARS = NULL)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{group_VARS}{Attributes in colData that define groups to apply the
normalization to.}
}
\value{
A ChIPtsne2_no_rowRanges object. where values have been scaled per region where max
equals 1.
}
\description{
Applies a normalization where the maximum value at each region is scaled to
\enumerate{
\item By default this is applied across all samples but you may supply one or
more variables from column/sample metadata to group them by.
}
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
# plotting raw values, it can be hard to assess differences between cells at
# regions with varying average enrichment.
plotSignalHeatmap(ct2)

# with region normalizatoin, it's much clearer that MCF10CA1 has lower
# enrichment
plotSignalHeatmap(normalizeSignalPerRegion(ct2))

# this data isn't complicated enough to warrant it, but normalization can be
# applied with sample groups
plotSignalHeatmap(normalizeSignalPerRegion(ct2, group_VARS = "cell"))
}
