% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce_plot_points.R
\name{plotDimReducePoints}
\alias{plotDimReducePoints}
\alias{plotDimReducePoints,ChIPtsne2_no_rowRanges-method}
\title{plotDimReducePoints}
\usage{
plotDimReducePoints(
  ct2,
  color_VAR = NULL,
  label_VAR = NULL,
  label_FUN = geom_label,
  label_size = 10,
  point_size = NULL,
  point_color_limits = c(NA, NA),
  has_symmetrical_limits = NULL,
  point_colors = NULL,
  extra_VARS = NULL,
  background_annotation_color = NULL,
  underlayer_FUN = function(p, ...) p,
  return_data = FALSE
)

\S4method{plotDimReducePoints}{ChIPtsne2_no_rowRanges}(
  ct2,
  color_VAR = NULL,
  label_VAR = NULL,
  label_FUN = geom_label,
  label_size = 10,
  point_size = NULL,
  point_color_limits = c(NA, NA),
  has_symmetrical_limits = NULL,
  point_colors = NULL,
  extra_VARS = NULL,
  background_annotation_color = NULL,
  underlayer_FUN = function(p, ...) p,
  return_data = FALSE
)
}
\arguments{
\item{ct2}{valid ChIPtsne2 after dimReduce has been run}

\item{color_VAR}{Control color assignment in plot. Can match entries in
\emph{either} sample metadata (colnames) or region metadata (rowRanges). Default
of NULL will plot max signal for all sample profiles. NA will perform no
color mapping.}

\item{label_VAR}{Categorical variable to label the mean position of. Use with \code{label_FUN} to control geom used.}

\item{label_FUN}{Function to add labels to plot. Only used when \code{label_VAR} is specified. Should be equivalent to geom_label: geom_text, ggrepel::geom_text_repel, or ggrepel::geom_label_repel. Must accept parameters, data, mapping, and show.legend.}

\item{label_size}{Font size of label.}

\item{point_size}{Size of points in plot.}

\item{point_color_limits}{color scale limits for continuous color_VAR.}

\item{has_symmetrical_limits}{If TRUE color scale limits will extend to equal
magnitude in positive and negative direction. Default is TRUE when negative
values are present and FALSE otherwise.}

\item{point_colors}{Either a vectors of colors to pass to
\link[ggplot2:scale_gradient]{ggplot2::scale_colour_gradientn} for continuous data, or a named vector of
colors for categorical data.}

\item{extra_VARS}{Extra attributes from rowData or colData to carry through. These attributes will be available for facetting and similar ggplot2 operations.}

\item{background_annotation_color}{Color to use for points not in facet. Default of NULL will not draw any background points.}

\item{underlayer_FUN}{Function to add to layer below background annotation.}

\item{return_data}{If TRUE, no plot is returned but the final plotted data.frame is returned instead. Default is FALSE.}
}
\value{
ggplot
}
\description{
plotDimReducePoints
}
\examples{
library(ggplot2)
ct2 = exampleChIPtsne2.with_meta() \%>\%
   dimReduceUMAP() \%>\%
   groupRegionsByDimReduceCluster(group_VAR = "umap_cluster", nearest_neighbors = 20) \%>\%
   groupRegionsBySignalCluster(group_VAR = "signal_cluster")

# default is max signal value per sample
plotDimReducePoints(ct2)
plotDimReducePoints(ct2, point_colors = c("gray", "red"))

#NA disable color
plotDimReducePoints(ct2, color_VAR = NA)

# color scale is different when negative values are present
ct2_diff = subsetSamples(ct2, cell == "MCF10A") -
  subsetSamples(ct2, cell == "MCF10AT1")
plotDimReducePoints(ct2_diff)

# categorical variables
plotDimReducePoints(ct2, "umap_cluster")
plotDimReducePoints(ct2, "umap_cluster", label_VAR = "umap_cluster")
#a named vector of colors for point_colors
plotDimReducePoints(ct2, "umap_cluster",
  point_colors = seqsetvis::safeBrew(as.character(1:4), "paired"))
plotDimReducePoints(ct2, c("umap_cluster", "signal_cluster"))
plotDimReducePoints(ct2, c("MCF10A_CTCF", "MCF10AT1_CTCF"))

# layer plot elements beneath the final plot with a function like this:
base_plot = function(p, xy_df, point_size, background_annotation_color){
  p +
    annotate("rect",
             xmin = 0, xmax = .3,
             ymin = -.05, ymax = .13,
             fill = "lightblue", color = "red")
}

plotDimReducePoints(ct2, extra_VARS = "peak_MCF10CA1_CTCF",
  background_annotation_color = "gray50",
  underlayer_FUN = base_plot) +
    facet_grid(peak_MCF10CA1_CTCF~name) +
    labs(title = "signal per sample facetted by peak_MCF10CA1_CTCF")
}
