% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsneList.R
\name{rbind,ChIPtsne2List-method}
\alias{rbind,ChIPtsne2List-method}
\title{rbind-ChIPtsne2List}
\usage{
\S4method{rbind}{ChIPtsne2List}(..., deparse.level = 1)
}
\arguments{
\item{...}{A single A ChIPtsne2List object.}

\item{deparse.level}{Not used.}
}
\value{
a ChIPtsne2 object of concatenated rows/regions of all items in input ChIPtsne2List
}
\description{
rbind-ChIPtsne2List
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2.by_peak = split(ct2, "peak_MCF10AT1_CTCF")
ct2.rbind = rbind(ct2.by_peak)
}
