% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce.R
\name{reportDimReduceMethod}
\alias{reportDimReduceMethod}
\title{reportDimReduceMethod}
\usage{
reportDimReduceMethod(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}
}
\value{
Character describing active dimReduction method.
}
\description{
reportDimReduceMethod
}
\examples{
ct2 = exampleChIPtsne2()
reportDimReduceMethod(ct2)
ct2 = dimReduceTSNE(ct2, perplexity = 25)
reportDimReduceMethod(ct2)
}
