% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_history.R
\name{rerun_history}
\alias{rerun_history}
\title{rerun_history}
\usage{
rerun_history(ct2, history_source)
}
\arguments{
\item{ct2}{A ChIPtsne2 object to apply history too.}

\item{history_source}{list of history items or a ChIPtsne2 object to pull history from.}
}
\value{
Input ct2 with all functions from history_source applied.
}
\description{
rerun_history
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2.c = centerProfilesAndTrim(ct2, view_size = 2e3)
ChIPtsne2.history(ct2.c)
#histories can be rerun just by supplying a source ct2 object
ct2.c_re1 = rerun_history(ct2, history_source = ct2.c)
ChIPtsne2.history(ct2.c_re1)
hist_src = ChIPtsne2.history(ct2.c)
#you can also supply the history items list, potentially modifying entries like so
hist_src$centerProfilesAndTrim$ARG$view_size = 1000
ct2.c_re2 = rerun_history(ct2, history_source = hist_src)
ChIPtsne2.history(ct2.c_re2)
}
