% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_ChIPtsne2_shiny}
\alias{run_ChIPtsne2_shiny}
\title{run_ChIPtsne2_shiny}
\usage{
run_ChIPtsne2_shiny(ct2)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}
}
\value{
ChIPtsne2 object modified with user's interactive annotations.
}
\description{
run_ChIPtsne2_shiny
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = dimReduceTSNE(ct2)
\dontrun{
  plotDimReducePoints(ct2, color_VAR = "peak_MCF10A_CTCF")
  ct2.ann = run_ChIPtsne2_shiny(ct2)
  plotDimReducePoints(ct2.ann, color_VAR = "peak_MCF10A_CTCF")
}


}
