% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FetchConfig.R
\name{runFetchAtRegions}
\alias{runFetchAtRegions}
\title{runFetchAtRegions}
\usage{
runFetchAtRegions(fetch_config, query_gr, use_cache = FALSE)
}
\arguments{
\item{fetch_config}{A FetchConfig object}

\item{query_gr}{A GRanges to fetch data for}

\item{use_cache}{If TRUE, default \link[BiocFileCache:BiocFileCache-class]{BiocFileCache::BiocFileCache} will be used. If FALSE, no caching will be done. You may also supply a user created \link[BiocFileCache:BiocFileCache-class]{BiocFileCache::BiocFileCache}.}
}
\value{
A list of 2 items prof_dt and query_gr.  prof_dt is a tidy data.table
of signal profiles.  query_gr is a GRanges that may have been modified from
input query_gr if signal profiles are flipped or centered according to
center_signal_at_max or flip_signal_mode in the signal config.
}
\description{
runFetchAtRegions
}
