% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{scale_color_config}
\alias{scale_color_config}
\title{scale_color_config}
\usage{
scale_color_config(object)
}
\arguments{
\item{object}{a Config object}
}
\value{
a ggplot2 scale_fill_manual
}
\description{
scale_color_config
}
