% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_getset.R
\name{setRegionMetaData}
\alias{setRegionMetaData}
\title{setRegionMetaData}
\usage{
setRegionMetaData(ct2, new_meta, silent = FALSE)
}
\arguments{
\item{ct2}{A ChIPtsne object}

\item{new_meta}{A data.frame with new metadata information. Must include same name_VAR as ct2 or have equivalent rownames. Variables already present in ct2 will result in overiting those variables.}
}
\value{
A modified ChIPtsne2 object with added/overwritten sample metadata.
}
\description{
setRegionMetaData
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
new_meta = getRegionMetaData(ct2)
new_meta[["10A_and_AT1"]] = ifelse(new_meta$peak_MCF10A_CTCF & new_meta$peak_MCF10AT1_CTCF, "yes", "no")
ct2 = setRegionMetaData(ct2, new_meta)
getRegionMetaData(ct2)

ct2.nrr = exampleChIPtsne2.with_meta()
rowRanges(ct2.nrr) = NULL
getRegionMetaData(ct2.nrr)
ct2.nrr = setRegionMetaData(ct2.nrr, new_meta)
getRegionMetaData(ct2.nrr)
}
