% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_getset.R
\name{setSampleMetaData}
\alias{setSampleMetaData}
\title{setSampleMetaData}
\usage{
setSampleMetaData(ct2, new_meta, silent = FALSE)
}
\arguments{
\item{ct2}{A ChIPtsne object}

\item{new_meta}{A data.frame with new metadata information. Must include same name_VAR as ct2 or have equivalent rownames. Variables already present in ct2 will result in overiting those variables.}

\item{silent}{If TRUE, no messages will be displayed. Default is FALSE.}
}
\value{
A modified ChIPtsne2 object with added/overwritten sample metadata.
}
\description{
setSampleMetaData
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
getSampleMetaData(ct2)
new_meta = data.frame(name = c("MCF10A_CTCF", "MCF10AT1_CTCF", "MCF10CA1_CTCF"), id1 = seq(3))
ct2 = setSampleMetaData(ct2, new_meta)
getSampleMetaData(ct2)

#metadata may be overriden
new_meta2 = data.frame(id1 = LETTERS[seq(3)], id2 = LETTERS[seq(3)])
rownames(new_meta2) = c("MCF10A_CTCF", "MCF10AT1_CTCF", "MCF10CA1_CTCF")
ct2 = setSampleMetaData(ct2, new_meta2)
getSampleMetaData(ct2)
}
