% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_seed.R
\name{setSeed}
\alias{setSeed}
\alias{setSeed,ChIPtsne2_no_rowRanges-method}
\title{setSeed}
\usage{
setSeed(ct2, seed = NULL)

\S4method{setSeed}{ChIPtsne2_no_rowRanges}(ct2, seed = NULL)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{seed}{New set.seed value. Default of NULL will use Sys.time. This will still be saved in history.}
}
\value{
A ChIPtsne2_no_rowRanges object. with seed stored in histroy
}
\description{
Calls \link{set.seed}.
}
\details{
Recorded in history to allow reproducibilty with \link{rerurun_history}
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = setSeed(ct2)
ChIPtsne2.history(ct2)$setSeed$ARG$seed
}
