% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_no_rowRanges.R
\name{show,ChIPtsne2_no_rowRanges-method}
\alias{show,ChIPtsne2_no_rowRanges-method}
\title{show_ChIPtsne2_no_rowRanges}
\usage{
\S4method{show}{ChIPtsne2_no_rowRanges}(object)
}
\arguments{
\item{object}{A ChIPtsne2_no_rowRanges object}
}
\description{
used by show
}
\examples{
ct2 = exampleChIPtsne2()
ct2
}
