% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FetchConfig.R
\name{show,FetchConfig-method}
\alias{show,FetchConfig-method}
\title{show_Config}
\usage{
\S4method{show}{FetchConfig}(object)
}
\arguments{
\item{object}{A FetchConfig object}
}
\description{
used by show
}
\examples{
bam_config_file = exampleBamConfigFile()
fetch_config = FetchConfig.load_config(bam_config_file)
fetch_config
}
