% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{subsetRegions}
\alias{subsetRegions}
\title{subsetRegions}
\usage{
subsetRegions(ct2, expr)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{expr}{expression, indicating columns to select from region/row metadata.}
}
\value{
A subsetted A ChIPtsne2_no_rowRanges object.
}
\description{
subsetRegions
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
subsetRegions(ct2, peak_MCF10AT1_CTCF == TRUE)
}
