% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_transform_signal.R
\docType{methods}
\name{transformSignal-ChIPtsne2_no_rowRanges}
\alias{transformSignal-ChIPtsne2_no_rowRanges}
\alias{transformSignal,ChIPtsne2_no_rowRanges-method}
\title{transformSignal method for objects of class \code{ChIPtsne2_no_rowRanges}.}
\usage{
\S4method{transformSignal}{ChIPtsne2_no_rowRanges}(ct2, transform_FUN)
}
\arguments{
\item{x}{A \code{ChIPtsne2_no_rowRanges} object.}
}
\description{
transformSignal method for objects of class \code{ChIPtsne2_no_rowRanges}.
}
