% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{translateCT2toSSV}
\alias{translateCT2toSSV}
\title{translateCT2toSSV}
\usage{
translateCT2toSSV(prof_dt)
}
\arguments{
\item{prof_dt}{A data.table generated by \emph{chiptsne2} from getTidyProfile with
default VAR names.}
}
\value{
The input \code{prof_dt} with colnames updated to be compatible with
\emph{seqsetvis}.
}
\description{
\emph{seqsetvis} creates data.table with certain default variables that
\emph{chiptsne2} uses more user-friendly values for. This function checks for
\emph{chiptsne2} defaults and converts them to \emph{seqsetvis}.
}
\examples{
prof_dt = exampleProfDT()
prof_dt_for_ct2 = translateSSVtoCT2(prof_dt)
ct2 = ChIPtsne2.from_tidy(prof_dt_for_ct2, exampleQueryGR())

prof_dt_ssv_again = translateCT2toSSV(prof_dt_for_ct2)
setequal(colnames(prof_dt), colnames(prof_dt_ssv_again))
}
