% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.R
\name{addRegionAnnotation}
\alias{addRegionAnnotation}
\title{addRegionAnnotation}
\usage{
addRegionAnnotation(
  ct2,
  anno_gr,
  anno_VAR = "annotation",
  anno_VAR_renames = anno_VAR,
  no_overlap_value = "no_hit",
  overlap_value = "hit"
)
}
\arguments{
\item{ct2}{A ChIPtsne object}

\item{anno_gr}{A GenomicRanges object to annotate ct2 based on overlap with rowRanges of ct2.}

\item{anno_VAR}{Attribute in mcols of anno_gr to pull values from.}

\item{anno_VAR_renames}{Matched vector to anno_VAR specificying final names in rowRanges of ct2. Essentially renames anno_VAR.}

\item{no_overlap_value}{Value for when there is no overlap with anno_gr. Default is "no_hit".}

\item{overlap_value}{Value for when there is an overlap, only relevant if anno_VAR is not in mcols of anno_gr. I.e. adding a single "hit" "no hit" annotation.}
}
\value{
A modified ChIPtsne2 object with added/overwritten sample metadata.
}
\description{
An alternative to setRegionMetaData that adds region metadata information based on overlaps with supplied anno_gr.
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
anno_gr = rowRanges(ct2)
anno_gr = subset(anno_gr, seqnames == "chr1")
anno_gr$start_pos = start(anno_gr)
anno_gr$end_pos = end(anno_gr)
anno_gr$chr_name = as.character(seqnames(anno_gr))
ct2.anno = addRegionAnnotation(ct2, anno_gr, anno_VAR = c("start_pos", "end_pos", "chr_name"))
rowRanges(ct2.anno)
}
