% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aggregate.R
\name{aggregateByGroup}
\alias{aggregateByGroup}
\title{aggregateByGroup}
\usage{
aggregateByGroup(
  ct2,
  group_VARS,
  sample_sep = " ",
  region_sep = " ",
  summary_FUN = mean
)
}
\arguments{
\item{ct2}{A ChIPtsne2 object.}

\item{group_VARS}{Attribute name to aggregate samples and/or regions to.  There will be 1 meta-sample per unique sampled entry in \code{group_VARS} and 1 meta-region per unique region entry. \code{group_VARS} may specify multiple attributes, in which case there will 1 meta-region and/or sample per combination of entries in all \code{group_VARS}.}

\item{sample_sep}{Separator character to use when pasting multiple sample grouping variables.}

\item{region_sep}{Separator character to use when pasting multiple region grouping variables.}

\item{summary_FUN}{function applied to every row and column group to summarize. Default is mean.}
}
\value{
Either a ChIPtsne2 or ChIPtsne2_no_rowRanges object, with meta-regions for combinations of \code{group_VARS} values if region aggregation occurred.
}
\description{
aggregateByGroup
}
\examples{
ct2 = exampleChIPtsne2.with_meta()

aggregateRegionsByGroup(ct2, c("peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"))
aggregateSamplesByGroup(ct2, c("mark"))
aggregateByGroup(ct2, c("mark", "peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"))

ct2.r1 = exampleChIPtsne2.with_meta()
colData(ct2.r1)$rep = "rep1"
colnames(ct2.r1) = paste0(colnames(ct2.r1), "_rep1")
ct2.r2 = exampleChIPtsne2.with_meta()
colData(ct2.r2)$rep = "rep2"
colnames(ct2.r2) = paste0(colnames(ct2.r2), "_rep2")
ct2.reps = cbind(ct2.r1, ct2.r2)
ct2.agg3 = aggregateByGroup(ct2.reps, c("cell", "mark", "peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"))
colData(ct2.agg3)
rowData(ct2.agg3)
}
