% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aggregate.R
\name{aggregateByGroup}
\alias{aggregateByGroup}
\title{aggregateByGroup}
\usage{
aggregateByGroup(ct2, group_VAR)
}
\arguments{
\item{ct2}{A ChIPtsne2 object.}

\item{group_VAR}{Attribute name to aggregate samples and/or regions to.  There will be 1 meta-sample per unique sampled entry in \code{group_VAR} and 1 meta-region per unique region entry. \code{group_VAR} may specify multiple attributes, in which case there will 1 meta-region and/or sample per combination of entries in all \code{group_VAR}.}
}
\value{
Either a ChIPtsne2 or ChIPtsne2_no_rowRanges object, with meta-regions for combinations of \code{group_VAR} values if region aggregation occurred.
}
\description{
aggregateByGroup
}
\examples{
ct2 = exampleChIPtsne2.with_meta()

aggregateRegionsByGroup(ct2, c("peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"))
aggregateSamplesByGroup(ct2, c("mark"))
aggregateByGroup(ct2, c("mark", "peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"))

ct2.r1 = exampleChIPtsne2.with_meta()
colData(ct2.r1)$rep = "rep1"
colnames(ct2.r1) = paste0(colnames(ct2.r1), "_rep1")
ct2.r2 = exampleChIPtsne2.with_meta()
colData(ct2.r2)$rep = "rep2"
colnames(ct2.r2) = paste0(colnames(ct2.r2), "_rep2")
ct2.reps = cbind(ct2.r1, ct2.r2)
aggregateByGroup(ct2.reps, c("cell", "mark", "peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"))
}
