% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aggregate.R
\name{aggregateRegionsByGroup}
\alias{aggregateRegionsByGroup}
\title{aggregateRegionsByGroup}
\usage{
aggregateRegionsByGroup(
  ct2,
  group_VAR,
  new_meta_VAR = ifelse(length(group_VAR) == 1, group_VAR, "meta_id")
)
}
\arguments{
\item{ct2}{A ChIPtsne2 object.}

\item{group_VAR}{Attribute name to aggregate regions to.  There will be 1 meta-region per unique entry in \code{group_VAR}. \code{group_VAR} may specify multiple attributes, in which case there will 1 meta-region per combination of entries in all \code{group_VAR}.}

\item{new_meta_VAR}{The new region variable of the resulting ChIPtsne2 object. Defaults to \code{group_VAR} for single \code{group_VAR} but then defaults to "meta_id" if there are multiple.}
}
\value{
A ChIPtsne2_no_rowRanges object with meta-regions for combinations of \code{group_VAR} values.
}
\description{
aggregateRegionsByGroup
}
\examples{
ct2 = exampleChIPtsne2.with_meta()

aggregateRegionsByGroup(ct2, "peak_MCF10A_CTCF")

aggregateRegionsByGroup(
  ct2,
  c("peak_MCF10A_CTCF", "peak_MCF10AT1_CTCF"),
  new_meta_VAR = "peak_overlap"
)
}
