% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot_lineplot.R
\name{aggregateRegionsByGroup}
\alias{aggregateRegionsByGroup}
\title{Title}
\usage{
aggregateRegionsByGroup(
  ct2,
  group_VAR,
  new_meta_VAR = ifelse(length(group_VAR) == 1, group_VAR, "meta_id")
)
}
\arguments{
\item{new_meta_VAR}{}
}
\description{
Title
}
