% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aggregate.R
\name{aggregateSamplesByGroup}
\alias{aggregateSamplesByGroup}
\title{aggregateSamplesByGroup}
\usage{
aggregateSamplesByGroup(
  ct2,
  group_VAR,
  new_meta_VAR = ifelse(length(group_VAR) == 1, group_VAR, "meta_name")
)
}
\arguments{
\item{ct2}{A ChIPtsne2 object.}

\item{group_VAR}{Attribute name to aggregate samples to.  There will be 1 meta-sample per unique entry in \code{group_VAR}. \code{group_VAR} may specify multiple attributes, in which case there will 1 meta-region per combination of entries in all \code{group_VAR}.}

\item{new_meta_VAR}{The new name variable of the resulting ChIPtsne2 object. Defaults to \code{group_VAR} for single \code{group_VAR} but then defaults to "meta_id" if there are multiple.}
}
\value{
A ChIPtsne2_no_rowRanges object with meta-regions for combinations of \code{group_VAR} values.
}
\description{
aggregateSamplesByGroup
}
\examples{
ct2 = exampleChIPtsne2.with_meta()

aggregateSamplesByGroup(ct2, "mark")

ct2.r1 = exampleChIPtsne2.with_meta()
colData(ct2.r1)$rep = "rep1"
colnames(ct2.r1) = paste0(colnames(ct2.r1), "_rep1")
ct2.r2 = exampleChIPtsne2.with_meta()
colData(ct2.r2)$rep = "rep2"
colnames(ct2.r2) = paste0(colnames(ct2.r2), "_rep2")
ct2.reps = cbind(ct2.r1, ct2.r2)
aggregateSamplesByGroup(ct2.reps, c("cell", "mark"), "name")
aggregateSamplesByGroup(ct2.reps, c("cell", "mark"))
}
