% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot_lineplot.R
\name{aggregateSamplesByGroup}
\alias{aggregateSamplesByGroup}
\title{aggregateSamplesByGroup}
\usage{
aggregateSamplesByGroup(
  ct2,
  group_VAR,
  new_meta_VAR = ifelse(length(group_VAR) == 1, group_VAR, "meta_name")
)
}
\arguments{
\item{new_meta_VAR}{}
}
\description{
aggregateSamplesByGroup
}
