% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_centroid_classifier.R
\name{calculateGroupCentroid}
\alias{calculateGroupCentroid}
\title{calculateGroupCentroid}
\usage{
calculateGroupCentroid(ct2, group_VARS)
}
\arguments{
\item{group_VARS}{}
}
\description{
calculateGroupCentroid
}
