% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsneList.R
\name{cbind,ChIPtsne2List-method}
\alias{cbind,ChIPtsne2List-method}
\title{cbind-ChIPtsne2List}
\usage{
\S4method{cbind}{ChIPtsne2List}(..., deparse.level = 1)
}
\arguments{
\item{ChIPtsne2List}{A ChIPtsne2List object.}
}
\value{
a ChIPtsne2 object of concatenated columns/samples of all items in input ChIPtsne2List
}
\description{
cbind-ChIPtsne2List
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2.by_cell = split(ct2, "cell")
ct2.cbind = cbind(ct2.by_cell)
}
