% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.alt_constructors.R
\name{ChIPtsne2.from_FetchConfig}
\alias{ChIPtsne2.from_FetchConfig}
\title{ChIPtsne2.from_FetchConfig}
\usage{
ChIPtsne2.from_FetchConfig(
  fetch_config,
  query_gr,
  region_metadata = NULL,
  obj_history = list(),
  init = TRUE
)
}
\arguments{
\item{fetch_config}{An object of class FetchConfig}

\item{query_gr}{The query GRanges object used to fetch prof_dt.}

\item{region_metadata}{Metadata to append to rowRanges, mcols of query_gr will also be used.}

\item{obj_history}{Existing history for object, may be used to describe origin on prof_dt or query_gr.}

\item{init}{If TRUE, initialize history with birthday, session_info, and chiptsne2_version}
}
\value{
A ChIPtsne2 created using profiles fetched used provided FetchConfig
}
\description{
ChIPtsne2.from_FetchConfig
}
\examples{
bam_cfg_f = system.file("extdata/bam_config.csv", package = "chiptsne2", mustWork = TRUE)
fetch_config = FetchConfig.load_config(bam_cfg_f)
query_gr = seqsetvis::CTCF_in_10a_overlaps_gr
ct2 = ChIPtsne2.from_FetchConfig(fetch_config, query_gr)
ct2
}
