% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_alt_constructors.R
\name{ChIPtsne2.from_tidy}
\alias{ChIPtsne2.from_tidy}
\title{ChIPtsne2.from_tidy}
\usage{
ChIPtsne2.from_tidy(
  prof_dt,
  query_gr,
  sample_metadata = NULL,
  region_metadata = NULL,
  name_VAR = "name",
  position_VAR = "position",
  value_VAR = "signal",
  region_VAR = "region",
  auto_sample_metadata = TRUE,
  obj_history = list(),
  fetch_config = FetchConfig.null(),
  init = TRUE
)
}
\arguments{
\item{prof_dt}{Profile data.table, as returned from seqsetvis::ssvFetch* functions}

\item{query_gr}{The query GRanges object used to fetch prof_dt.}

\item{sample_metadata}{Metadata for entries in prof_dt's name_VAR, must include name_VAR}

\item{region_metadata}{Metadata to append to rowRanges, mcols of query_gr will also be used.}

\item{name_VAR}{Variable name that contains sample ids/names. Links prof_dt to meta_dt. Default is "name".}

\item{position_VAR}{Variable name that contains positional information in prof_dt. Default is "position".}

\item{value_VAR}{Variable name that contains signal value information in prof_dt. Default is "signal".}

\item{region_VAR}{Variable name that contains region ID information in prof_dt. Default is "region".}

\item{auto_sample_metadata}{If true, additional attributes in prof_dt will used for metadata (minus certain region related attributes such as seqnames, start, end, etc.)}

\item{obj_history}{Existing history for object, may be used to describe origin on prof_dt or query_gr.}

\item{fetch_config}{A FetchConfig object. Optional. Allows signal to be fetched as needed.}

\item{init}{If TRUE, initialize history with birthday, session_info, and chiptsne2_version}
}
\value{
ChIPtsne2 object with tsne results
}
\description{
ChIPtsne2.from_tidy
}
\examples{
library(seqsetvis)
data(CTCF_in_10a_overlaps_gr)
query_gr = CTCF_in_10a_overlaps_gr
data(CTCF_in_10a_profiles_dt)
prof_dt = CTCF_in_10a_profiles_dt
ct2 = ChIPtsne2.from_tidy(prof_dt, query_gr)
ct2
}
