% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.alt_constructors.R
\name{ChIPtsne2.from_tidy}
\alias{ChIPtsne2.from_tidy}
\title{ChIPtsne2.from_tidy}
\usage{
ChIPtsne2.from_tidy(
  prof_dt,
  query_gr,
  meta_dt = NULL,
  name_VAR = "sample",
  position_VAR = "x",
  value_VAR = "y",
  region_VAR = "id"
)
}
\arguments{
\item{prof_dt}{Profile data.table, as returned from seqsetvis::ssvFetch* functions}

\item{query_gr}{The query GRanges object used to fetch prof_dt.}

\item{meta_dt}{Additional information for each name_VAR entry in prof_dt.}

\item{name_VAR}{Variable name that contains sample ids/names. Links prof_dt to meta_dt. Default is "sample".}

\item{position_VAR}{Variable name that contains positional information in prof_dt. Default is "x".}

\item{value_VAR}{Variable name that contains signal value information in prof_dt. Default is "y".}

\item{region_VAR}{Variable name that contains region ID information in prof_dt. Default is "id".}
}
\value{
ChIPtsne2 object with tsne results
}
\description{
ChIPtsne2.from_tidy
}
\examples{
query_gr = CTCF_in_10a_overlaps_gr
prof_dt = CTCF_in_10a_profiles_dt
ChIPtsne2.from_tidy(prof_dt, query_gr)
}
