% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.alt_constructors.R
\name{ChIPtsne2.history}
\alias{ChIPtsne2.history}
\title{ChIPtsne2.history}
\usage{
ChIPtsne2.history(ct2)
}
\arguments{
\item{ct2}{}
}
\value{
list of history items
}
\description{
ChIPtsne2.history
}
