% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.R
\name{ChIPtsne2}
\alias{ChIPtsne2}
\title{ChIPtsne2}
\usage{
ChIPtsne2(
  rowToRowMat = matrix(0, 0, 0),
  colToRowMatCols = list(),
  name_VAR = "name",
  position_VAR = "position",
  value_VAR = value,
  region_VAR = "region",
  fetch_config = FetchConfig.null(),
  ...
)
}
\arguments{
\item{rowToRowMat}{Value for rowToRowMat slot. A matrix with profile data where colnames equals items in colToRowMatCols. The rownames are the rownames/region ids.}

\item{colToRowMatCols}{Value for colToRowMatCols slot. A named list where names equals colnames of x and list values are colnames of rowToRowMat.}

\item{name_VAR}{Name variable, a single character.}

\item{position_VAR}{Position variable, a single character.}

\item{value_VAR}{Value variable, a single character.}

\item{region_VAR}{Region variable, a single character.}

\item{fetch_config}{A \link{FetchConfig} object.}

\item{...}{Arguments passed to \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}. Must include: rowRanges, colData, assays, metadata}
}
\description{
This is not the recommended function for user generation of a ChIPtsne2
object. Users should use either \link{ChIPtsne2.from_FetchConfig} or
\link{ChIPtsne2.from_tidy} in the majority of cases.
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
# really, never make the ChIPtsne2 object this way.
# if you must, you may inspect the individual elements below:
ct2_new = ChIPtsne2(
    rowToRowMat = rowToRowMat(ct2),
    colToRowMatCols = colToRowMatCols(ct2),
    name_VAR = getNameVariable(ct2),
    position_VAR = getPositionVariable(ct2),
    value_VAR = getValueVariable(ct2),
    region_VAR = getRegionVariable(ct2),
    fetch_config = FetchConfig.null(),
    rowRanges = rowRanges(ct2),
    colData = colData(ct2),
    assays = assays(ct2),
    metadata = ChIPtsne2.history(ct2)
)
ct2_new


}
