% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.R
\name{ChIPtsne2}
\alias{ChIPtsne2}
\alias{rowRanges<-,ChIPtsne2,NULL-method}
\title{ChIPtsne2}
\usage{
ChIPtsne2(
  rowToRowMat = matrix(0, 0, 0),
  colToRowMatCols = list(),
  name_VAR = "sample",
  position_VAR = "x",
  value_VAR = "y",
  region_VAR = "id",
  fetch_config = FetchConfig.null(),
  ...
)

\S4method{rowRanges}{ChIPtsne2,NULL}(x, ...) <- value
}
\description{
ChIPtsne2

rowRanges
}
