% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_no_rowRanges.R
\name{ChIPtsne2_no_rowRanges}
\alias{ChIPtsne2_no_rowRanges}
\title{ChIPtsne2_no_rowRanges}
\usage{
ChIPtsne2_no_rowRanges(
  rowToRowMat = matrix(0, 0, 0),
  colToRowMatCols = list(),
  name_VAR = "sample",
  position_VAR = "x",
  value_VAR = "y",
  region_VAR = "id",
  fetch_config = FetchConfig.null(),
  ...
)
}
\description{
This is the base class for ChIPtsne2. Both classes support \code{rowData} but only
ChIPtsne2 provides \code{rowRanges} and is therefore suitable for \code{GenomicRanges}
functions.
}
\details{
ChIPtsne2 object may be converted to ChIPtsne2_no_rowRanges simply by NULL
assignment to rowRagnes. i.e. rowRanges(ct2) = NULL.
}
