% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_center_signal.R
\name{centerProfilesAndTrim}
\alias{centerProfilesAndTrim}
\alias{centerProfilesAndTrim,ChIPtsne2_no_rowRanges,numeric-method}
\title{centerProfilesAndTrim}
\usage{
centerProfilesAndTrim(ct2, view_size)

\S4method{centerProfilesAndTrim}{ChIPtsne2_no_rowRanges,numeric}(ct2, view_size)
}
\arguments{
\item{ct2}{A ChIPtsne2 object}

\item{view_size}{bp range to search for max}
}
\value{
A chiptsne2 object updated to reflect centering procedure. Some x
values will have been lost.
}
\description{
centerProfilesAndTrim
}
\examples{
library(magrittr)
query_gr = exampleQueryGR()
query_gr = seqsetvis::prepare_fetch_GRanges_width(query_gr, win_size = 50)
prof_dt = exampleProfDT()
meta_dt = prof_dt \%>\%
  dplyr::select(sample) \%>\%
  unique \%>\%
  tidyr::separate(sample, c("cell", "mark"), remove = FALSE)
ct2 = ChIPtsne2.from_tidy(prof_dt, query_gr, sample_metadata = meta_dt)
ct2.c = centerProfilesAndTrim(ct2, view_size = 500)
}
