% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_alt_constructors.R
\name{cloneChIPtsne2}
\alias{cloneChIPtsne2}
\alias{cloneChIPtsne2,ChIPtsne2_no_rowRanges-method}
\alias{cloneChIPtsne2_fromTidy}
\alias{cloneChIPtsne2_fromTidy,ChIPtsne2_no_rowRanges-method}
\title{cloneChIPtsne2}
\usage{
cloneChIPtsne2(
  ct2,
  new_rowToRowMat = NULL,
  new_colToRowMatCols = NULL,
  new_name_VAR = NULL,
  new_position_VAR = NULL,
  new_value_VAR = NULL,
  new_region_VAR = NULL,
  new_fetch_config = NULL,
  new_rowRanges = NULL,
  new_colData = NULL,
  new_assays = NULL,
  new_metadata = NULL
)

\S4method{cloneChIPtsne2}{ChIPtsne2_no_rowRanges}(
  ct2,
  new_rowToRowMat = NULL,
  new_colToRowMatCols = NULL,
  new_name_VAR = NULL,
  new_position_VAR = NULL,
  new_value_VAR = NULL,
  new_region_VAR = NULL,
  new_fetch_config = NULL,
  new_rowRanges = NULL,
  new_colData = NULL,
  new_assays = NULL,
  new_metadata = NULL
)

cloneChIPtsne2_fromTidy(
  ct2,
  new_prof_dt = NULL,
  new_obj_history = NULL,
  new_name_VAR = NULL,
  new_position_VAR = NULL,
  new_value_VAR = NULL,
  new_region_VAR = NULL,
  new_fetch_config = NULL,
  new_query_gr = NULL,
  init = FALSE,
  new_sample_metadata = NULL,
  new_region_metadata = NULL
)

\S4method{cloneChIPtsne2_fromTidy}{ChIPtsne2_no_rowRanges}(
  ct2,
  new_prof_dt = NULL,
  new_obj_history = NULL,
  new_name_VAR = NULL,
  new_position_VAR = NULL,
  new_value_VAR = NULL,
  new_region_VAR = NULL,
  new_fetch_config = NULL,
  new_query_gr = NULL,
  init = FALSE,
  new_sample_metadata = NULL,
  new_region_metadata = NULL
)
}
\arguments{
\item{ct2}{A ChIPtsne2 object. or A ChIPtsne2_no_rowRanges object.}

\item{new_rowToRowMat}{rowToRowMat slot override}

\item{new_colToRowMatCols}{colToRowMatCols slot override}

\item{new_name_VAR}{name_VAR slot override}

\item{new_position_VAR}{position_VAR slot override}

\item{new_value_VAR}{value_VAR slot override}

\item{new_region_VAR}{region_VAR slot override}

\item{new_fetch_config}{slot override}

\item{new_rowRanges}{rowRanges slot override}

\item{new_colData}{colData slot override}

\item{new_assays}{assays slot override}

\item{new_metadata}{metadata slot override}
}
\value{
Clone of input A ChIPtsne2 object. or A ChIPtsne2_no_rowRanges object. with specified slots modified.
}
\description{
Create a clone (copy) of a ChIPtsne2 object replacing any specified components.
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
getNameVariable(ct2)
ct2 = cloneChIPtsne2(ct2, new_name_VAR = "new_name")
getNameVariable(ct2)


}
