% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_alt_constructors.R
\name{cloneChIPtsne2_fromTidy}
\alias{cloneChIPtsne2_fromTidy}
\alias{cloneChIPtsne2_fromTidy,ChIPtsne2_no_rowRanges-method}
\title{cloneChIPtsne2_fromTidy}
\usage{
cloneChIPtsne2_fromTidy(
  ct2,
  new_prof_dt = NULL,
  new_name_VAR = NULL,
  new_position_VAR = NULL,
  new_value_VAR = NULL,
  new_region_VAR = NULL,
  new_fetch_config = NULL,
  new_rowRanges = NULL,
  new_sample_metadata = NULL,
  new_region_metadata = NULL,
  new_obj_history = NULL
)

\S4method{cloneChIPtsne2_fromTidy}{ChIPtsne2_no_rowRanges}(
  ct2,
  new_prof_dt = NULL,
  new_name_VAR = NULL,
  new_position_VAR = NULL,
  new_value_VAR = NULL,
  new_region_VAR = NULL,
  new_fetch_config = NULL,
  new_rowRanges = NULL,
  new_sample_metadata = NULL,
  new_region_metadata = NULL,
  new_obj_history = NULL
)
}
\arguments{
\item{ct2}{A ChIPtsne2 object. or A ChIPtsne2_no_rowRanges object.}

\item{new_prof_dt}{tidy representation of rowToRowMat and colToRowMatCols.
Could be modified output of \link{getTidyProfile}.}

\item{new_name_VAR}{name_VAR slot override}

\item{new_position_VAR}{position_VAR slot override}

\item{new_value_VAR}{value_VAR slot override}

\item{new_region_VAR}{region_VAR slot override}

\item{new_fetch_config}{slot override}

\item{new_rowRanges}{rowRanges slot override}

\item{new_sample_metadata}{Could be modified output of \link{getSampleMetaData} to
override colData}

\item{new_region_metadata}{Could be modified output of \link{getRegionMetaData} to
override rowData}

\item{new_obj_history}{object history, aka metadata, slot override}
}
\value{
Clone of input A ChIPtsne2 object. or A ChIPtsne2_no_rowRanges object. with specified
slots modified.
}
\description{
cloneChIPtsne2_fromTidy
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
plotSignalLinePlot(ct2)

# one applicatation of cloning would be to extract tidy profiles,
# modify somehow, and then reinsert.
prof_dt = getTidyProfile(ct2)
prof_dt[name == "MCF10A_CTCF", value := value/4]
ct2.clone1 = cloneChIPtsne2_fromTidy(ct2, new_prof_dt = prof_dt)
plotSignalLinePlot(ct2.clone1)
}
