% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_flip_signal.R
\name{flipProfilesToMatch}
\alias{flipProfilesToMatch}
\alias{flipProfilesToMatch,ChIPtsne2_no_rowRanges-method}
\title{flipProfilesToMatch}
\usage{
flipProfilesToMatch(ct2, highest_on_right = TRUE)

\S4method{flipProfilesToMatch}{ChIPtsne2_no_rowRanges}(ct2, highest_on_right = FALSE)
}
\arguments{
\item{ct2}{A ChIPtsne2 object}
}
\value{
A chiptsne2 object updated such that signal "tilts" in the same way.
When signal is flipped, strand of rowRanges is set to negative.
}
\description{
For unstranded data, a peak with higher signal on one side is not
meaningfully different from its mirror image. This procedure identifies
"tilted" peaks and flips those with more signal right-of-center. By flipping,
rowRanges becomes stranded, flipped regions get assigned (-) strand with
uaffected regions getting (+) strand
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2_left = flipProfilesToMatch(ct2)
ct2_right = flipProfilesToMatch(ct2, highest_on_right = TRUE)

prof_original = getTidyProfile(ct2)
prof_original$group = "original"

prof_left = getTidyProfile(ct2_left)
prof_left$group = "flip_left"

prof_right = getTidyProfile(ct2_right)
prof_right$group = "flip_right"

prof_dt = rbind(prof_left, prof_right, prof_original)
prof_dt = prof_dt[, list(y = mean(y)), .(x, group)]

library(ggplot2)
ggplot(prof_dt, aes(x = x, y = y, color = group)) + geom_path()
}
