% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/class_ChIPtsneList.R,
%   R/class_ChIPtsne_getset.R, R/class_ChIPtsne_no_rowRanges.R
\name{getset functions}
\alias{getset functions}
\alias{setNameVariable}
\alias{swapNameVariable}
\alias{getNameVariable}
\alias{setValueVariable}
\alias{getValueVariable}
\alias{setRegionVariable}
\alias{getRegionVariable}
\alias{setPositionVariable}
\alias{getPositionVariable}
\alias{getRegionMetaData}
\alias{getMaxValueData}
\alias{setNameVariable,ChIPtsne2List-method}
\alias{swapNameVariable,ChIPtsne2List-method}
\alias{getNameVariable,ChIPtsne2List-method}
\alias{setValueVariable,ChIPtsne2List-method}
\alias{getValueVariable,ChIPtsne2List-method}
\alias{setRegionVariable,ChIPtsne2List-method}
\alias{getRegionVariable,ChIPtsne2List-method}
\alias{setPositionVariable,ChIPtsne2List-method}
\alias{getPositionVariable,ChIPtsne2List-method}
\alias{setNameVariable,ChIPtsne2_no_rowRanges-method}
\alias{swapNameVariable,ChIPtsne2_no_rowRanges-method}
\alias{getNameVariable,ChIPtsne2_no_rowRanges-method}
\alias{setValueVariable,ChIPtsne2_no_rowRanges-method}
\alias{getValueVariable,ChIPtsne2_no_rowRanges-method}
\alias{setRegionVariable,ChIPtsne2_no_rowRanges-method}
\alias{getRegionVariable,ChIPtsne2_no_rowRanges-method}
\alias{setPositionVariable,ChIPtsne2_no_rowRanges-method}
\alias{getPositionVariable,ChIPtsne2_no_rowRanges-method}
\alias{getSampleMetaData}
\alias{rowToRowMat}
\alias{rowToRowMat,ChIPtsne2_no_rowRanges-method}
\alias{colToRowMatCols}
\alias{colToRowMatCols,ChIPtsne2_no_rowRanges-method}
\alias{rowToRowMat<-}
\alias{rowToRowMat<-,ChIPtsne2_no_rowRanges-method}
\alias{colToRowMatCols<-}
\alias{colToRowMatCols<-,ChIPtsne2_no_rowRanges-method}
\title{getSampleMetaData}
\usage{
setNameVariable(ct2, new_VAR)

swapNameVariable(ct2, new_VAR)

getNameVariable(ct2)

setValueVariable(ct2, new_VAR)

getValueVariable(ct2)

setRegionVariable(ct2, new_VAR)

getRegionVariable(ct2)

setPositionVariable(ct2, new_VAR)

getPositionVariable(ct2)

getRegionMetaData(ct2, select_VARS = NULL, include_value_max = FALSE)

getMaxValueData(ct2)

\S4method{setNameVariable}{ChIPtsne2List}(ct2, new_VAR)

\S4method{swapNameVariable}{ChIPtsne2List}(ct2, new_VAR)

\S4method{getNameVariable}{ChIPtsne2List}(ct2)

\S4method{setValueVariable}{ChIPtsne2List}(ct2, new_VAR)

\S4method{getValueVariable}{ChIPtsne2List}(ct2)

\S4method{setRegionVariable}{ChIPtsne2List}(ct2, new_VAR)

\S4method{getRegionVariable}{ChIPtsne2List}(ct2)

\S4method{setPositionVariable}{ChIPtsne2List}(ct2, new_VAR)

\S4method{getPositionVariable}{ChIPtsne2List}(ct2)

\S4method{setNameVariable}{ChIPtsne2_no_rowRanges}(ct2, new_VAR)

\S4method{swapNameVariable}{ChIPtsne2_no_rowRanges}(ct2, new_VAR)

\S4method{getNameVariable}{ChIPtsne2_no_rowRanges}(ct2)

\S4method{setValueVariable}{ChIPtsne2_no_rowRanges}(ct2, new_VAR)

\S4method{getValueVariable}{ChIPtsne2_no_rowRanges}(ct2)

\S4method{setRegionVariable}{ChIPtsne2_no_rowRanges}(ct2, new_VAR)

\S4method{getRegionVariable}{ChIPtsne2_no_rowRanges}(ct2)

\S4method{setPositionVariable}{ChIPtsne2_no_rowRanges}(ct2, new_VAR)

\S4method{getPositionVariable}{ChIPtsne2_no_rowRanges}(ct2)

getSampleMetaData(ct2, select_VARS = NULL)

rowToRowMat(x)

\S4method{rowToRowMat}{ChIPtsne2_no_rowRanges}(x)

colToRowMatCols(x)

\S4method{colToRowMatCols}{ChIPtsne2_no_rowRanges}(x)

rowToRowMat(x) <- value

\S4method{rowToRowMat}{ChIPtsne2_no_rowRanges}(x) <- value

colToRowMatCols(x) <- value

\S4method{colToRowMatCols}{ChIPtsne2_no_rowRanges}(x) <- value
}
\arguments{
\item{ct2}{A ChIPtsne object}

\item{new_VAR}{New variable name. If it is already present this is an error.
Use \link{swapNameVariable} if you intend to use an existing metadata variable
as the new name variable.}

\item{select_VARS}{character vector of variables to select from sample
metadata. Default of NULL will select all available sample metadata
variables.}

\item{x}{A ChIPtsne2_no_rowRanges object.}

\item{value}{New value for colToRowMatCols slot. A named list where names equals colnames of x and list values are colnames of rowToRowMat.}
}
\value{
data.frame with sample meta data, similar to colData but suitable for
tidyverse operations.
}
\description{
getters and setters for ChIPtsne2 and ChIPtsne2_no_rowRanges objects

Contains the signal for all samples. Each row is a region and blocks of
columns per sample are defined in \code{colToRowMatCols}.

maps ChIPtsne2 column names to columns in rowToRowMat
}
\examples{
ct2 = exampleChIPtsne2()
getSampleMetaData(ct2)
getSampleMetaData(ct2, "sample")
}
