% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/class_ChIPtsneList.R,
%   R/class_ChIPtsne_getset.R, R/class_ChIPtsne_no_rowRanges.R
\name{setNameVariable}
\alias{setNameVariable}
\alias{swapNameVariable}
\alias{getNameVariable}
\alias{setValueVariable}
\alias{getValueVariable}
\alias{setRegionVariable}
\alias{getRegionVariable}
\alias{setPositionVariable}
\alias{getPositionVariable}
\alias{getRegionMetaData}
\alias{setNameVariable,ChIPtsne2List-method}
\alias{swapNameVariable,ChIPtsne2List-method}
\alias{getNameVariable,ChIPtsne2List-method}
\alias{setValueVariable,ChIPtsne2List-method}
\alias{getValueVariable,ChIPtsne2List-method}
\alias{setRegionVariable,ChIPtsne2List-method}
\alias{getRegionVariable,ChIPtsne2List-method}
\alias{setPositionVariable,ChIPtsne2List-method}
\alias{getPositionVariable,ChIPtsne2List-method}
\alias{setNameVariable,ChIPtsne2_no_rowRanges-method}
\alias{swapNameVariable,ChIPtsne2_no_rowRanges-method}
\alias{getNameVariable,ChIPtsne2_no_rowRanges-method}
\alias{setValueVariable,ChIPtsne2_no_rowRanges-method}
\alias{getValueVariable,ChIPtsne2_no_rowRanges-method}
\alias{setRegionVariable,ChIPtsne2_no_rowRanges-method}
\alias{getRegionVariable,ChIPtsne2_no_rowRanges-method}
\alias{setPositionVariable,ChIPtsne2_no_rowRanges-method}
\alias{getPositionVariable,ChIPtsne2_no_rowRanges-method}
\alias{rowToRowMat}
\alias{rowToRowMat,ChIPtsne2_no_rowRanges-method}
\alias{colToRowMatCols}
\alias{colToRowMatCols,ChIPtsne2_no_rowRanges-method}
\alias{colToRowMatCols<-}
\alias{colToRowMatCols<-,ChIPtsne2_no_rowRanges-method}
\title{getters and setters for ChIPtsne2 and ChIPtsne2_no_rowRanges objects}
\usage{
setNameVariable(ct2, new_name_VAR)

swapNameVariable(ct2, new_name_VAR)

getNameVariable(ct2)

setValueVariable(ct2, new_value_VAR)

getValueVariable(ct2)

setRegionVariable(ct2, new_region_VAR)

getRegionVariable(ct2)

setPositionVariable(ct2, new_position_VAR)

getPositionVariable(ct2)

getRegionMetaData(ct2, select_VARS = NULL)

\S4method{setNameVariable}{ChIPtsne2List}(ct2, new_name_VAR)

\S4method{swapNameVariable}{ChIPtsne2List}(ct2, new_name_VAR)

\S4method{getNameVariable}{ChIPtsne2List}(ct2)

\S4method{setValueVariable}{ChIPtsne2List}(ct2, new_value_VAR)

\S4method{getValueVariable}{ChIPtsne2List}(ct2)

\S4method{setRegionVariable}{ChIPtsne2List}(ct2, new_region_VAR)

\S4method{getRegionVariable}{ChIPtsne2List}(ct2)

\S4method{setPositionVariable}{ChIPtsne2List}(ct2, new_position_VAR)

\S4method{getPositionVariable}{ChIPtsne2List}(ct2)

\S4method{setNameVariable}{ChIPtsne2_no_rowRanges}(ct2, new_name_VAR)

\S4method{swapNameVariable}{ChIPtsne2_no_rowRanges}(ct2, new_name_VAR)

\S4method{getNameVariable}{ChIPtsne2_no_rowRanges}(ct2)

\S4method{setValueVariable}{ChIPtsne2_no_rowRanges}(ct2, new_value_VAR)

\S4method{getValueVariable}{ChIPtsne2_no_rowRanges}(ct2)

\S4method{setRegionVariable}{ChIPtsne2_no_rowRanges}(ct2, new_region_VAR)

\S4method{getRegionVariable}{ChIPtsne2_no_rowRanges}(ct2)

\S4method{setPositionVariable}{ChIPtsne2_no_rowRanges}(ct2, new_position_VAR)

\S4method{getPositionVariable}{ChIPtsne2_no_rowRanges}(ct2)

rowToRowMat(x)

\S4method{rowToRowMat}{ChIPtsne2_no_rowRanges}(x)

colToRowMatCols(x)

\S4method{colToRowMatCols}{ChIPtsne2_no_rowRanges}(x)

colToRowMatCols(x) <- value

\S4method{colToRowMatCols}{ChIPtsne2_no_rowRanges}(x) <- value
}
\arguments{
\item{x}{A ChIPtsne2_no_rowRanges object.}

\item{value}{New value for colToRowMatCols slot. A named list where names equals colnames of x and list values are colnames of rowToRowMat.}
}
\description{
Contains the signal for all samples. Each row is a region and blocks of
columns per sample are defined in \code{colToRowMatCols}.

maps ChIPtsne2 column names to columns in rowToRowMat
}
