% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{setNameVariable}
\alias{setNameVariable}
\alias{swapNameVariable}
\alias{getNameVariable}
\alias{setValueVariable}
\alias{getValueVariable}
\alias{setRegionVariable}
\alias{getRegionVariable}
\alias{setPositionVariable}
\alias{getPositionVariable}
\alias{getRegionMetaData}
\title{getters and setters for ChIPtsne2 and ChIPtsne2_no_rowRanges objects}
\usage{
setNameVariable(ct2, new_name_VAR)

swapNameVariable(ct2, new_name_VAR)

getNameVariable(ct2)

setValueVariable(ct2, new_value_VAR)

getValueVariable(ct2)

setRegionVariable(ct2, new_region_VAR)

getRegionVariable(ct2)

setPositionVariable(ct2, new_position_VAR)

getPositionVariable(ct2)

getRegionMetaData(ct2, select_VARS = NULL)
}
\description{
getters and setters for ChIPtsne2 and ChIPtsne2_no_rowRanges objects
}
