% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_grouping.R
\name{groupRegionsManually}
\alias{groupRegionsManually}
\alias{groupRegionsManually,ChIPtsne2_no_rowRanges-method}
\alias{groupRegionsByValues}
\alias{groupRegionsByValues,ChIPtsne2_no_rowRanges-method}
\title{groupRegionsManually}
\usage{
groupRegionsManually(ct2, assignment, group_VAR = "group_id")

\S4method{groupRegionsManually}{ChIPtsne2_no_rowRanges}(ct2, assignment, group_VAR = "group_id")

groupRegionsByValues(
  ct2,
  value_test,
  yes = NULL,
  no = NULL,
  group_VAR = "value_id"
)

\S4method{groupRegionsByValues}{ChIPtsne2_no_rowRanges}(
  ct2,
  value_test,
  yes = NULL,
  no = NULL,
  group_VAR = "value_id"
)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{assignment}{A data.frame or named character vector. See details.}

\item{group_VAR}{Attribute name to add to rowData for storing new row/region grouping assignment. Will be overwritten if it exists.}

\item{value_test}{Expression to evaluate on values}

\item{yes}{Value to assign to group_VAR where expression is TRUE. Defaults to "yes"}

\item{no}{Value to assign to group_VAR where expression is FALSE. Defaults to existing values for group_VAR when group_VAR exists, "no" if not.}
}
\value{
Updated ChIPtsne2 object with group assignment added to rowData using specified group_VAR.

Updated ChIPtsne2 object with group assignment added to rowData using specified group_VAR.
}
\description{
groupRegionsManually

groupRegionsByValues
}
\examples{
ct2 = exampleChIPtsne2()

# data.frame assignment
np_grs = examplePeaks()
memb_grs = seqsetvis::ssvOverlapIntervalSets(np_grs)
assign_df = seqsetvis::ssvFactorizeMembTable(memb_grs)
assign_df$not_used = "A"
assign_df$group2 = "B"
ct2.grouped = groupRegionsManually(ct2, assign_df, group_VAR = c("group", "group2"))
rowData(ct2.grouped)

# character vector assignment
manual_groups = assign_df$group
names(manual_groups) = assign_df$id
ct2.grouped2 = groupRegionsManually(ct2, manual_groups, group_VAR = c("group_by_vector"))
rowData(ct2.grouped2)
ct2 = exampleChIPtsne2.with_meta()
ct2.high_10a = subsetValues(ct2, MCF10A_CTCF > 30)
nrow(ct2.high_10a)

ct2_exp = groupRegionsByValues(ct2, MCF10A_CTCF > 30)
table(rowData(ct2_exp)$value_id)
#character version of expression is OK for programming purposes
ct2_str = groupRegionsByValues(ct2, "MCF10A_CTCF > 30")
table(rowData(ct2_str)$value_id)

min_value = 30
ct2_grp = ct2
ct2_grp = groupRegionsByValues(ct2_grp, MCF10A_CTCF > min_value, yes = "high_10a", no = "bg")
# don't include `no` after initializing group variable
ct2_grp = groupRegionsByValues(ct2_grp, MCF10AT1_CTCF > min_value, yes = "high_at1")
ct2_grp = groupRegionsByValues(ct2_grp, MCF10AT1_CTCF > min_value & MCF10A_CTCF > min_value, yes = "high_at1_and_10a")
table(rowData(ct2_grp)$value_id)

#and since it's a grouping variable, we can use it in plots
plotSignalHeatmap(ct2_grp, group_VARS = "value_id")
}
