% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_grouping.R
\name{groupRegionsManually}
\alias{groupRegionsManually}
\alias{groupRegionsManually,ChIPtsne2_no_rowRanges-method}
\title{groupRegionsManually}
\usage{
groupRegionsManually(ct2, assignment, group_VAR = "group_id")

\S4method{groupRegionsManually}{ChIPtsne2_no_rowRanges}(ct2, assignment, group_VAR = "group_id")
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{assignment}{A data.frame or named character vector. See details.}

\item{group_VAR}{Attribute name to add to rowData for storing new row/region grouping assignment. Will be overwritten if it exists.

Assignment may be either a data.frame or named character vector.

If data.frame, must have rownames setequal to rownames of \code{ct2} or contain the same region variable as \code{ct2}. Multiple \code{group_VAR} may be specified and all must be present in data.frame.

If a named character vector, names must be setequal to rownames of \code{ct2}. Only a signle \code{group_VAR} is allowed.}
}
\value{
Updated ChIPtsne2 object with group assignment added to rowData using specified group_VAR.
}
\description{
groupRegionsManually
}
\examples{
ct2 = exampleChIPtsne2()

# data.frame assignment
np_grs = examplePeaks()
memb_grs = seqsetvis::ssvOverlapIntervalSets(np_grs)
assign_df = seqsetvis::ssvFactorizeMembTable(memb_grs)
assign_df$not_used = "A"
assign_df$group2 = "B"
ct2.grouped = groupRegionsManually(ct2, assign_df, group_VAR = c("group", "group2"))
rowData(ct2.grouped)

# character vector assignment
manual_groups = assign_df$group
names(manual_groups) = assign_df$id
ct2.grouped2 = groupRegionsManually(ct2, manual_groups, group_VAR = c("group_by_vector"))
rowData(ct2.grouped2)
}
