% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_arithmetic.R
\name{apply_ChIPtsne2_operator}
\alias{apply_ChIPtsne2_operator}
\alias{apply_ChIPtsne2_operator.num}
\alias{-,ChIPtsne2_no_rowRanges,ChIPtsne2_no_rowRanges-method}
\alias{+,ChIPtsne2_no_rowRanges,ChIPtsne2_no_rowRanges-method}
\alias{/,ChIPtsne2_no_rowRanges,ChIPtsne2_no_rowRanges-method}
\alias{*,ChIPtsne2_no_rowRanges,ChIPtsne2_no_rowRanges-method}
\alias{-,ChIPtsne2_no_rowRanges,numeric-method}
\alias{+,ChIPtsne2_no_rowRanges,numeric-method}
\alias{/,ChIPtsne2_no_rowRanges,numeric-method}
\alias{*,ChIPtsne2_no_rowRanges,numeric-method}
\alias{-,numeric,ChIPtsne2_no_rowRanges-method}
\alias{+,numeric,ChIPtsne2_no_rowRanges-method}
\alias{/,numeric,ChIPtsne2_no_rowRanges-method}
\alias{*,numeric,ChIPtsne2_no_rowRanges-method}
\title{ChipTSNE2 arithmetic
Arithmetic operators are supports for ChIPtsne2 and ChIPtsne2_no_rowRanges objects.}
\usage{
apply_ChIPtsne2_operator(e1, e2, operator = "-")

apply_ChIPtsne2_operator.num(e1, e2, operator = "-")

\S4method{-}{ChIPtsne2_no_rowRanges,ChIPtsne2_no_rowRanges}(e1, e2)

\S4method{+}{ChIPtsne2_no_rowRanges,ChIPtsne2_no_rowRanges}(e1, e2)

\S4method{/}{ChIPtsne2_no_rowRanges,ChIPtsne2_no_rowRanges}(e1, e2)

\S4method{*}{ChIPtsne2_no_rowRanges,ChIPtsne2_no_rowRanges}(e1, e2)

\S4method{-}{ChIPtsne2_no_rowRanges,numeric}(e1, e2)

\S4method{+}{ChIPtsne2_no_rowRanges,numeric}(e1, e2)

\S4method{/}{ChIPtsne2_no_rowRanges,numeric}(e1, e2)

\S4method{*}{ChIPtsne2_no_rowRanges,numeric}(e1, e2)

\S4method{-}{numeric,ChIPtsne2_no_rowRanges}(e1, e2)

\S4method{+}{numeric,ChIPtsne2_no_rowRanges}(e1, e2)

\S4method{/}{numeric,ChIPtsne2_no_rowRanges}(e1, e2)

\S4method{*}{numeric,ChIPtsne2_no_rowRanges}(e1, e2)
}
\description{
+, -, *, and /
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = ct2[, c("MCF10A_CTCF", "MCF10AT1_CTCF")]
ct2.by_cell = split(ct2, "cell")
plotSignalLinePlot(ct2)

#operators can be used on 2 ChIPtsne2_no_rowRanges objects or 1 object and a number
ct2.diff = ct2.by_cell$MCF10A - ct2.by_cell$MCF10AT1
plotSignalLinePlot(ct2.diff)

# numeric operators work in either order
ct2.2div10a = cbind(2 / (ct2.by_cell$MCF10A+1), ct2.by_cell$MCF10AT1)
plotSignalLinePlot(ct2.2div10a)
ct2.10adiv2 = cbind(ct2.by_cell$MCF10A / 2, ct2.by_cell$MCF10AT1)
plotSignalLinePlot(ct2.10adiv2)

}
