% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_grouping.R
\name{sortRegions}
\alias{sortRegions}
\alias{sortRegions,ChIPtsne2_no_rowRanges-method}
\title{sortRegions}
\usage{
sortRegions(ct2, group_VAR = NULL)

\S4method{sortRegions}{ChIPtsne2_no_rowRanges}(ct2, group_VAR = NULL)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{group_VAR}{Attribute name to add to rowData for storing new row/region grouping assignment. Will be overwritten if it exists.}
}
\value{
A ChIPtsne2_no_rowRanges object. with rows sorted by \code{group_VAR} and within groups by signal decreasing.
}
\description{
sortRegions
}
\examples{
ct2 = exampleChIPtsne2()
rowData(ct2)
ct2.sorted = sortRegions(ct2, group_VAR = "peak_MCF10CA1_CTCF")
rowData(ct2.sorted)

ct2.sorted2 = sortRegions(ct2, group_VAR = c("peak_MCF10AT1_CTCF", "peak_MCF10CA1_CTCF"))
rowData(ct2.sorted2)
}
