% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_transform_signal.R
\docType{methods}
\name{transformSignal}
\alias{transformSignal}
\alias{transformSignal-ChIPtsne2_no_rowRanges}
\alias{transformSignal,ChIPtsne2_no_rowRanges-method}
\title{transformSignal method for objects of class \code{ChIPtsne2_no_rowRanges}.}
\usage{
transformSignal(ct2, transform_FUN)

\S4method{transformSignal}{ChIPtsne2_no_rowRanges}(ct2, transform_FUN)
}
\arguments{
\item{x}{A \code{ChIPtsne2_no_rowRanges} object.}
}
\description{
transformSignal method for objects of class \code{ChIPtsne2_no_rowRanges}.
}
\examples{
library(ggplot2)
ct2 = exampleChIPtsne2.with_meta()
plotSignalLinePlot(ct2) + expand_limits(y = 30)
# a transformation function takes a matrix and returns the same size matrix
trans_fun_1 = function(mat, ...){
    mat/2
}

ct2.t1 = transformSignal(ct2, trans_fun_1)
plotSignalLinePlot(ct2.t1) + expand_limits(y = 30)

# as a second argument, sample metadata is passed in per sample
trans_fun_2 = function(mat, sample_meta){
  if(sample_meta$cell == "MCF10CA1"){
    mat / 4
  }else{
    mat
  }
}
ct2.t2 = transformSignal(ct2, trans_fun_2)
plotSignalLinePlot(ct2.t2) + expand_limits(y = 30)

}
