% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce.R
\name{.pca_from_profile_mat}
\alias{.pca_from_profile_mat}
\title{.pca_from_profile_mat}
\usage{
.pca_from_profile_mat(prof_mat)
}
\arguments{
\item{prof_mat}{matrix of profile data}
}
\value{
data.frame of reduced dimensions by PCA
}
\description{
.pca_from_profile_mat
}
\examples{
ct2 = exampleChIPtsne2()
prof_mat = rowToRowMat(ct2)
.pca_from_profile_mat(prof_mat)
}
