% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dim_reduce_plots.R
\name{.plotDimReducePoints}
\alias{.plotDimReducePoints}
\title{plotDimReducePoints}
\usage{
.plotDimReducePoints(ct2, color_VAR = NULL, point_size = NULL)
}
\arguments{
\item{ct2}{valid ChIPtsne2 after dimReduce has been run}

\item{color_VAR}{Control color assignment in plot. Can match entries in *either* sample metadata (colnames) or region metadata (rowRanges). Default of NULL will plot max signal for all sample profiles. NA will perform no color mapping.}

\item{point_size}{Size of points in plot.}
}
\value{
ggplot
}
\description{
plotDimReducePoints
}
\examples{
ct2 = exampleChIPtsne2() \%>\%
   dimReduceUMAP() \%>\%
   groupRegionByDimReduceCluster(group_VAR = "umap_cluster") \%>\%
   groupRegionBySignalCluster(group_VAR = "signal_cluster")

plotDimReducePoints(ct2, NA)
plotDimReducePoints(ct2)
plotDimReducePoints(ct2, "umap_cluster")
plotDimReducePoints(ct2, c("umap_cluster", "signal_cluster"))
plotDimReducePoints(ct2, c("MCF10A_CTCF", "MCF10AT1_CTCF"))
}
