% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce.R
\name{.post_dim_reduce}
\alias{.post_dim_reduce}
\title{importFrom scales rescale}
\usage{
.post_dim_reduce(xy_df, prof_mat, norm1 = TRUE, high_topright = TRUE)
}
\description{
importFrom scales rescale
}
