% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{.save_config}
\alias{.save_config}
\title{internal function used by FetchConfig.save_config FetchConfigSignal.save_config and FetchConfigFeatures.save_config}
\usage{
.save_config(
  object,
  file,
  slots_to_save,
  kvp_slots,
  toss_names = "summary_FUN"
)
}
\description{
internal function used by FetchConfig.save_config FetchConfigSignal.save_config and FetchConfigFeatures.save_config
}
