% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{.shiny_secondary_plot}
\alias{.shiny_secondary_plot}
\title{.shiny_secondary_plot}
\usage{
.shiny_secondary_plot(plot_ct2, tx_df, plotType, selAttribute, selIds)
}
\description{
.shiny_secondary_plot
}
\examples{

plot_ct2 = exampleChIPtsne2.with_meta()
plot_ct2 = dimReduceTSNE(plot_ct2)
tx_df = getRegionMetaData(plot_ct2)
plotType = "line_color_selection"
selAttribute = "peak_MCF10A_CTCF"
selIds = character()
selIds = rownames(plot_ct2)[1:10]

.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "line_color_selection",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = rownames(plot_ct2)[1:10]
)
.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "line_color_selection",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = character()
)
.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "line_color_annotation",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = rownames(plot_ct2)[1:10]
)
.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "line_color_annotation",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = character()
)
.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "heatmap_cluster_selection",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = rownames(plot_ct2)[1:10]
)
.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "heatmap_cluster_selection",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = character()
)
.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "heatmap_cluster_annotation",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = rownames(plot_ct2)[1:10]
)
.shiny_secondary_plot(
  plot_ct2,
  tx_df,
  "heatmap_cluster_annotation",
  selAttribute = "peak_MCF10A_CTCF",
  selIds = character()
)

}
