% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{.show_Config}
\alias{.show_Config}
\title{.show_Config}
\usage{
.show_Config(qc)
}
\arguments{
\item{qc}{A Config object}
}
\description{
used by show
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
qc = ConfigFeatures.parse(feature_config_file)
qc
}
