% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{get_mapped_reads}
\alias{get_mapped_reads}
\title{get_mapped_reads}
\usage{
get_mapped_reads(bam_file)
}
\arguments{
\item{bam_file}{A bam file.  Matching .bai file must exist.}
}
\value{
The number of mapped reads in bam file.
}
\description{
get_mapped_reads
}
\examples{
data_dir = system.file("extdata", package = "seqsetvis", mustWork = TRUE)
bam_file = dir(data_dir, pattern = "bam$", full.names = TRUE)[1]

get_mapped_reads(bam_file)
}
