% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne.R
\name{getSampleMetaData}
\alias{getSampleMetaData}
\title{getSampleMetaData}
\usage{
getSampleMetaData(ct2, select_VARS = NULL)
}
\arguments{
\item{ct2}{A ChIPtsne object}

\item{select_VARS}{character vector of variables to select from sample
metadata. Default of NULL will select all available sample metadata
variables.}
}
\value{
data.frame with sample meta data, similar to colData but suitable for
  tidyverse operations.
}
\description{
getSampleMetaData
}
\examples{
ct2 = exampleChIPtsne2()
getSampleMetaData(ct2)
getSampleMetaData(ct2, "sample")
}
