% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_centroid_classifier.R
\name{groupRegionsByCentroidDistance}
\alias{groupRegionsByCentroidDistance}
\title{groupRegionsByCentroidDistance}
\usage{
groupRegionsByCentroidDistance(
  ct2,
  centroid,
  group_VAR,
  ambiguous_value = "ambiguous",
  tolerance = 0
)
}
\arguments{
\item{tolerance}{}
}
\description{
groupRegionsByCentroidDistance
}
