% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{mutateRegions}
\alias{mutateRegions}
\title{mutateRegions}
\usage{
mutateRegions(.data, ...)
}
\arguments{
\item{.data}{See \link[dplyr:mutate]{dplyr::mutate}}

\item{...}{Passed to \link[dplyr:mutate]{dplyr::mutate}}
}
\value{
A A ChIPtsne2_no_rowRanges object. with modified rowData/region metadata.
}
\description{
Should work just like \link[dplyr:mutate]{dplyr::mutate} but applied to rowData/region metadata of ChIPtsne2 objects.
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
rowData(ct2)
getRegionMetaData(ct2)
ct2 = mutateRegions(ct2, new_col = peak_MCF10A_CTCF | peak_MCF10AT1_CTCF)
rowData(ct2)
getRegionMetaData(ct2)
}
