% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{mutateSamples}
\alias{mutateSamples}
\title{mutateSamples}
\usage{
mutateSamples(.data, ...)
}
\arguments{
\item{.data}{See \link[dplyr:mutate]{dplyr::mutate}}

\item{...}{Passed to \link[dplyr:mutate]{dplyr::mutate}}
}
\value{
A A ChIPtsne2_no_rowRanges object. with modified colData/sample metadata.
}
\description{
Should work just like \link[dplyr:mutate]{dplyr::mutate} but applied to colData/sample metadata of ChIPtsne2 objects.
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
colData(ct2)
getSampleMetaData(ct2)
ct2 = mutateSamples(ct2, cell_mark = paste(cell, mark))
colData(ct2)
getSampleMetaData(ct2)
}
