% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce_plot_bins.R
\name{.plotDimReduceBins}
\alias{.plotDimReduceBins}
\title{plotDimReduceBins}
\usage{
.plotDimReduceBins(
  ct2,
  facet_rows = ct2@name_VAR,
  facet_columns = NULL,
  xmin = -Inf,
  xmax = Inf,
  agg_FUN = max,
  x_bins = NULL,
  y_bins = x_bins,
  bg_color = "gray60",
  min_size = 1,
  bin_fill_limits = c(NA, NA),
  has_symmetrical_limits = NULL,
  bin_colors = NULL,
  extra_VARS = character(),
  return_data = FALSE
)
}
\arguments{
\item{return_data}{}
}
\description{
plotDimReduceBins
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
ct2 = dimReducePCA(ct2)
plotDimReduceBins(ct2)
}
