% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce_plot_points.R
\name{plotDimReducePoints}
\alias{plotDimReducePoints}
\alias{plotDimReducePoints,ChIPtsne2_no_rowRanges-method}
\title{plotDimReducePoints}
\usage{
plotDimReducePoints(
  ct2,
  color_VAR = NULL,
  point_size = NULL,
  point_color_limits = c(NA, NA),
  has_symmetrical_limits = NULL,
  point_colors = NULL,
  extra_VARS = NULL,
  background_annotation_color = NULL,
  underlayer_FUN = function(p) p
)

\S4method{plotDimReducePoints}{ChIPtsne2_no_rowRanges}(
  ct2,
  color_VAR = NULL,
  point_size = NULL,
  point_color_limits = c(NA, NA),
  has_symmetrical_limits = NULL,
  point_colors = NULL,
  extra_VARS = NULL,
  background_annotation_color = NULL,
  underlayer_FUN = function(p) p
)
}
\arguments{
\item{ct2}{valid ChIPtsne2 after dimReduce has been run}

\item{color_VAR}{Control color assignment in plot. Can match entries in
\emph{either} sample metadata (colnames) or region metadata (rowRanges). Default
of NULL will plot max signal for all sample profiles. NA will perform no
color mapping.}

\item{point_size}{Size of points in plot.}

\item{point_color_limits}{color scale limits for continuous color_VAR.}

\item{has_symmetrical_limits}{If TRUE color scale limits will extend to equal
magnitude in positive and negative direction. Default is TRUE when negative
values are present and FALSE otherwise.}

\item{point_colors}{Either a vectors of colors to pass to
\link[ggplot2:scale_gradient]{ggplot2::scale_colour_gradientn} for continuous data, or a named vector of
colors for categorical data.}

\item{extra_VARS}{Extra attributes from rowData or colData to carry through. These attributes will be available for facetting and similar ggplot2 operations.}

\item{background_annotation_color}{Color to use for points not in facet. Default of NULL will not draw any background points.}

\item{underlayer_FUN}{Function to add to layer below background annotation.}
}
\value{
ggplot
}
\description{
plotDimReducePoints
}
