% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce_plot_points.R
\name{plotDimReducePoints}
\alias{plotDimReducePoints}
\alias{plotDimReducePoints,ChIPtsne2_no_rowRanges-method}
\title{plotDimReducePoints}
\usage{
plotDimReducePoints(
  ct2,
  color_VAR = NULL,
  point_size = NULL,
  point_color_limits = c(NA, NA),
  has_symmetrical_limits = NULL,
  point_colors = NULL,
  extra_VARS = NULL,
  background_annotation_color = NULL,
  underlayer_FUN = function(p) p
)

\S4method{plotDimReducePoints}{ChIPtsne2_no_rowRanges}(
  ct2,
  color_VAR = NULL,
  point_size = NULL,
  point_color_limits = c(NA, NA),
  has_symmetrical_limits = NULL,
  point_colors = NULL,
  extra_VARS = NULL,
  background_annotation_color = NULL,
  underlayer_FUN = function(p) p
)
}
\arguments{
\item{ct2}{valid ChIPtsne2 after dimReduce has been run}

\item{color_VAR}{Control color assignment in plot. Can match entries in
\emph{either} sample metadata (colnames) or region metadata (rowRanges). Default
of NULL will plot max signal for all sample profiles. NA will perform no
color mapping.}

\item{point_size}{Size of points in plot.}

\item{point_color_limits}{color scale limits for continuous color_VAR.}

\item{has_symmetrical_limits}{If TRUE color scale limits will extend to equal
magnitude in positive and negative direction. Default is TRUE when negative
values are present and FALSE otherwise.}

\item{point_colors}{Either a vectors of colors to pass to
\link[ggplot2:scale_gradient]{ggplot2::scale_colour_gradientn} for continuous data, or a named vector of
colors for categorical data.}

\item{extra_VARS}{Extra attributes from rowData or colData to carry through. These attributes will be available for facetting and similar ggplot2 operations.}

\item{background_annotation_color}{Color to use for points not in facet. Default of NULL will not draw any background points.}

\item{underlayer_FUN}{Function to add to layer below background annotation.}
}
\value{
ggplot
}
\description{
plotDimReducePoints
}
\examples{
library(ggplot2)
ct2 = exampleChIPtsne2.with_meta() \%>\%
   dimReduceUMAP() \%>\%
   groupRegionsByDimReduceCluster(group_VAR = "umap_cluster") \%>\%
   groupRegionsBySignalCluster(group_VAR = "signal_cluster")

# default is max signal value per sample
plotDimReducePoints(ct2)
plotDimReducePoints(ct2, point_colors = c("gray", "red"))

#NA disable color
plotDimReducePoints(ct2, color_VAR = NA)
ct2_diff = subsetSamples(ct2, cell == "MCF10A") -
  subsetSamples(ct2, cell == "MCF10AT1")
plotDimReducePoints(ct2_diff)
plotDimReducePoints(ct2, "umap_cluster")
#a named vector of colors for point_colors
plotDimReducePoints(ct2, "umap_cluster", point_colors = seqsetvis::safeBrew(as.character(1:4), "paired"))
plotDimReducePoints(ct2, c("umap_cluster", "signal_cluster"))
plotDimReducePoints(ct2, c("MCF10A_CTCF", "MCF10AT1_CTCF"))

# layer plot elements beneath the final plot with a function like this:
base_plot = function(p){
  p +
    annotate("rect",
             xmin = 0, xmax = .3,
             ymin = -.05, ymax = .13,
             fill = "gray80", color = "red")
}

plotDimReducePoints(ct2, extra_VARS = "peak_MCF10CA1_CTCF",
  background_annotation_color = "gray50",
  underlayer_FUN = base_plot) +
    facet_grid(peak_MCF10CA1_CTCF~sample) +
    labs(title = "signal per sample facetted by peak_MCF10CA1_CTCF")
}
