% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dim_reduce_plot_summary_profiles.R
\name{plotDimReduceSummaryProfiles}
\alias{plotDimReduceSummaryProfiles}
\title{plotDimReduceSummaryProfiles}
\usage{
plotDimReduceSummaryProfiles(
  ct2,
  color_VAR = ct2@name_VAR,
  x_bins = nrow(ct2)^0.5,
  y_bins = x_bins,
  extra_VARS = character(),
  xrng = NULL,
  yrng = NULL,
  value_limits = c(0, 1),
  ma_size = 2,
  n_splines = 10,
  p = NULL,
  N_floor = 0,
  N_ceiling = NULL,
  min_size = 0.3,
  return_data = FALSE
)
}
\arguments{
\item{return_data}{}
}
\description{
plotDimReduceSummaryProfiles
}
