% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{separateRegions}
\alias{separateRegions}
\title{separateRegions}
\usage{
separateRegions(
  data,
  col,
  into,
  sep = "[^[:alnum:]]+",
  remove = TRUE,
  convert = FALSE,
  extra = "warn",
  fill = "warn",
  ...
)
}
\arguments{
\item{data}{See \link[tidyr:separate]{tidyr::separate}}

\item{col}{See \link[tidyr:separate]{tidyr::separate}}

\item{into}{See \link[tidyr:separate]{tidyr::separate}}

\item{sep}{See \link[tidyr:separate]{tidyr::separate}}

\item{remove}{See \link[tidyr:separate]{tidyr::separate}}

\item{convert}{See \link[tidyr:separate]{tidyr::separate}}

\item{extra}{See \link[tidyr:separate]{tidyr::separate}}

\item{fill}{See \link[tidyr:separate]{tidyr::separate}}

\item{...}{Passed to \link[tidyr:separate]{tidyr::separate}}
}
\value{
A A ChIPtsne2_no_rowRanges object. with modified rowData/region metadata.
}
\description{
Should work just like \link[tidyr:separate]{tidyr::separate} but applied to rowData/region metadata of ChIPtsne2 objects.
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
rowData(ct2)
getRegionMetaData(ct2)
ct2 = mutateRegions(ct2, col_ab = paste("a", "b"))
ct2 = separateRegions(ct2, "col_ab", c("a", "b"), sep = " ", remove = FALSE)
rowData(ct2)
getRegionMetaData(ct2)
}
