% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{separateSamples}
\alias{separateSamples}
\title{separateSamples}
\usage{
separateSamples(
  data,
  col,
  into,
  sep = "[^[:alnum:]]+",
  remove = TRUE,
  convert = FALSE,
  extra = "warn",
  fill = "warn",
  ...
)
}
\arguments{
\item{data}{See \link[tidyr:separate]{tidyr::separate}}

\item{col}{See \link[tidyr:separate]{tidyr::separate}}

\item{into}{See \link[tidyr:separate]{tidyr::separate}}

\item{sep}{See \link[tidyr:separate]{tidyr::separate}}

\item{remove}{See \link[tidyr:separate]{tidyr::separate}}

\item{convert}{See \link[tidyr:separate]{tidyr::separate}}

\item{extra}{See \link[tidyr:separate]{tidyr::separate}}

\item{fill}{See \link[tidyr:separate]{tidyr::separate}}

\item{...}{Passed to \link[tidyr:separate]{tidyr::separate}}
}
\value{
A A ChIPtsne2_no_rowRanges object. with modified colData/sample metadata.
}
\description{
Should work just like \link[tidyr:separate]{tidyr::separate} but applied to colData/sample metadata of ChIPtsne2 objects.
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
colData(ct2)
getSampleMetaData(ct2)
ct2 = separateSamples(ct2, "cell_mark", c("v1", "v2"), sep = " ", remove = FALSE)
colData(ct2)
getSampleMetaData(ct2)
}
