% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_no_rowRanges.R
\name{split,ChIPtsne2_no_rowRanges,ANY-method}
\alias{split,ChIPtsne2_no_rowRanges,ANY-method}
\title{split-ChIPtsne2_no_rowRanges}
\usage{
\S4method{split}{ChIPtsne2_no_rowRanges,ANY}(x, f = NULL, drop = FALSE, ...)
}
\arguments{
\item{x}{A ChIPtsne2_no_rowRanges object.}

\item{f}{Either a row/region or column/sample specification in one of 2
modes. 1) a metadata attribute present in colData or rowData. Or 2) a
vector of the same length as rows or columns assigning groups to split into.}

\item{drop}{As in base split, logical indicating if levels that do not occur should be dropped (if f is a factor or a list).}

\item{...}{Not used. Will result in an error.}
}
\value{
A ChIPtsne2List where each entry is grouping value specified by \code{f}.
}
\description{
split-ChIPtsne2_no_rowRanges
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
split(ct2, "name")
split(ct2, c("cell", "mark"))
split(ct2, colnames(ct2))
split(ct2, "cell")
split(ct2, "peak_MCF10CA1_CTCF")
split(ct2, c("peak_MCF10AT1_CTCF", "peak_MCF10CA1_CTCF"))
split(ct2, ct2$cell)
split(ct2, paste(ct2$cell, ct2$mark, sep = "_"))

sample_meta_data = getSampleMetaData(ct2)
region_meta_data = getRegionMetaData(ct2)

split(ct2, sample_meta_data$mark)
split(ct2, region_meta_data$peak_MCF10A_CTCF)

}
