% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_no_rowRanges.R
\name{split,ChIPtsne2_no_rowRanges,ANY-method}
\alias{split,ChIPtsne2_no_rowRanges,ANY-method}
\title{split-ChIPtsne2_no_rowRanges}
\usage{
\S4method{split}{ChIPtsne2_no_rowRanges,ANY}(x, f = NULL, drop = FALSE, ...)
}
\arguments{
\item{ChIPtsne2_no_rowRanges}{}
}
\description{
split-ChIPtsne2_no_rowRanges
}
\examples{
split(ct2, "sample")
split(ct2, colnames(ct2))
split(ct2, "cell")
split(ct2, "peak_MCF10CA1_CTCF")
split(ct2, ct2$cell)

sample_meta_data = getSampleMetaData(ct2)
region_meta_data = getRegionMetaData(ct2)

split(ct2, sample_meta_data$mark)
split(ct2, region_meta_data$peak_MCF10A_CTCF)

}
