% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChIPtsne_no_rowRanges.R
\name{[,ChIPtsne2_no_rowRanges,ANY,ANY,ANY-method}
\alias{[,ChIPtsne2_no_rowRanges,ANY,ANY,ANY-method}
\title{ChIPtsne2_no_rowRanges array-like access}
\usage{
\S4method{[}{ChIPtsne2_no_rowRanges,ANY,ANY,ANY}(x, i, j, drop = TRUE)
}
\description{
ChIPtsne2_no_rowRanges array-like access
}
