% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{subsetCol}
\alias{subsetCol}
\title{subsetCol}
\usage{
subsetCol(ct2, subset_expression)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{value_test}{expression, indicating columns to select from sample/column metadata.}
}
\value{
A subsetted A ChIPtsne2_no_rowRanges object.
}
\description{
subsetCol
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
#deprecated, use subsetRegions instead
subsetSamples(ct2, cell \%in\% c("MCF10A", "MCF10AT1"))
}
