% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{subsetRow}
\alias{subsetRow}
\title{subsetRow}
\usage{
subsetRow(ct2, expr)
}
\arguments{
\item{ct2}{\code{doc_ct2_nrr()}}

\item{expr}{expression, indicating columns to select from region/row metadata.}
}
\value{
A subsetted \code{doc_ct2_nrr()}
}
\description{
subsetRow
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
#deprecated, use subsetRegions instead
subsetRegions(ct2, peak_MCF10AT1_CTCF == TRUE)
}
