% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{subsetSamples}
\alias{subsetSamples}
\title{subsetSamples}
\usage{
subsetSamples(ct2, expr)
}
\arguments{
\item{ct2}{\code{doc_ct2_nrr()}}

\item{expr}{expression, indicating columns to select from sample/column metadata.}
}
\value{
A subsetted \code{doc_ct2_nrr()}
}
\description{
subsetSamples
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
subsetSamples(ct2, cell \%in\% c("MCF10A", "MCF10AT1"))
}
