% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{subsetSamples}
\alias{subsetSamples}
\title{subsetSamples}
\usage{
subsetSamples(ct2, expr)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{expr}{expression, indicating columns to select from sample/column metadata.}
}
\value{
A subsetted A ChIPtsne2_no_rowRanges object.
}
\description{
subsetSamples
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
subsetSamples(ct2, cell \%in\% c("MCF10A", "MCF10AT1"))
}
