% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ChIPtsne_dplyr.R
\name{subsetValues}
\alias{subsetValues}
\title{subsetValues}
\usage{
subsetValues(ct2, value_test)
}
\arguments{
\item{ct2}{A ChIPtsne2_no_rowRanges object.}

\item{value_test}{expression, indicating column values to filter from assay slot.}
}
\value{
A subsetted A ChIPtsne2_no_rowRanges object.
}
\description{
subsetValues
}
\examples{
ct2 = exampleChIPtsne2.with_meta()
subsetValues(ct2, MCF10A_CTCF > 20 & MCF10AT1_CTCF > 20)
subsetValues(ct2, MCF10A_CTCF > Inf)

min_signal = 20
ct2 = subsetValues(ct2, MCF10A_CTCF > min_signal & MCF10AT1_CTCF > min_signal)
ct2
}
