% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_examples.R
\name{translateSSVtoCT2}
\alias{translateSSVtoCT2}
\title{translateSSVtoCT2}
\usage{
translateSSVtoCT2(prof_dt)
}
\arguments{
\item{prof_dt}{A data.table generated by \emph{seqsetvis} from ssvFetchGRanges}
}
\description{
\emph{seqsetvis} creates data.table with certain default variables that \emph{chiptsne2} uses more user-friendly values for. This function checks for \emph{seqsetvis} defaults and converts them to \emph{chiptsne2}.
}
\examples{
prof_dt = exampleProfDT()
prof_dt = translateSSVtoCT2(prof_dt)
ct2 = ChIPtsne2.from_tidy(prof_dt, exampleQueryGR())
}
